% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_serosurvey.R
\name{add_age_bins}
\alias{add_age_bins}
\title{Add bins based on age intervals.}
\usage{
add_age_bins(survey_features)
}
\arguments{
\item{survey_features}{A dataframe containing age_min and age_max columns representing
the minimum and maximum age boundaries for each group.}
}
\value{
A dataframe with an additional 'group' column representing the group interval
for each row based on the age_min and age_max columns.
}
\description{
It generates a new column 'group' in the survey_features dataframe, representing
the group interval for each row based on the age_min and age_max columns.
}
