% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_seromodel.R
\name{add_age_group_to_serosurvey}
\alias{add_age_group_to_serosurvey}
\title{Adds age group marker to serosurvey}
\usage{
add_age_group_to_serosurvey(serosurvey)
}
\arguments{
\item{serosurvey}{\describe{
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{age_min}}{Floor value of the average between age_min and age_max}
\item{\code{age_max}}{The size of the sample}
\item{\code{sample_size}}{Number of samples for each age group}
\item{\code{n_seropositive}}{Number of positive samples for each age group}
}}
}
\description{
Adds age group marker to serosurvey
}
