% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stan_data.R
\name{build_stan_data}
\alias{build_stan_data}
\title{Builds stan data for sampling depending on the selected model}
\usage{
build_stan_data(
  serosurvey,
  model_type = "constant",
  foi_prior = sf_uniform(),
  foi_index = NULL,
  is_seroreversion = FALSE,
  seroreversion_prior = sf_none()
)
}
\arguments{
\item{serosurvey}{\describe{
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{age_min}}{Floor value of the average between age_min and age_max}
\item{\code{age_max}}{The size of the sample}
\item{\code{sample_size}}{Number of samples for each age group}
\item{\code{n_seropositive}}{Number of positive samples for each age group}
}}

\item{model_type}{Type of the model. Either "constant", "age" or "time"}

\item{foi_prior}{Force-of-infection distribution specified by means of
the helper functions. Currently available options are:
\describe{
\item{\code{sf_normal}}
\item{\code{sf_uniform}}
\item{\code{sf_none}}
}}

\item{foi_index}{Integer vector specifying the age-groups for which
force-of-infection values will be estimated}

\item{is_seroreversion}{Boolean specifying whether to include
seroreversion rate estimation in the model}

\item{seroreversion_prior}{seroreversion distribution specified by means of
the helper functions. Currently available options are:
\describe{
\item{\code{sf_normal}}
\item{\code{sf_uniform}}
\item{\code{sf_none}}
}}
}
\value{
List with necessary data for sampling the specified model
}
\description{
Builds stan data for sampling depending on the selected model
}
