% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chagas2012.R
\docType{data}
\name{chagas2012}
\alias{chagas2012}
\title{Chagas seroprevalence data in serofoi}
\format{
\subsection{\code{chagas2012}}{

A \verb{<data.frame>} with 4 rows and 5 columns:
\describe{
\item{survey_year}{Year in which the serosurvey was conducted}
\item{n_sample}{Number of collected samples per age group}
\item{n_seropositive}{Number of positive samples per age group}
\item{age_min}{Age group minimal age}
\item{age_max}{Age group maximal age}
}
}
}
\usage{
data(chagas2012)
}
\description{
Datasets that measure the seroprevalence of IgG antibodies against
Trypanosoma cruzi infection in rural areas of Colombia corresponding
to a serosurvey conducted in 2012 for a rural indigenous community
known to have long-term endemic transmission, where some control
interventions have taken place over the years.
}
\examples{
data(chagas2012)
}
\keyword{datasets}
