% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_age_constraints}
\alias{check_age_constraints}
\title{Check min and max age consistency for validation purposes}
\usage{
check_age_constraints(survey_features)
}
\value{
Boolean checking consistency
}
\description{
Check min and max age consistency for validation purposes
}
\keyword{internal}
