% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{compare_and_save_best_model}
\alias{compare_and_save_best_model}
\title{Fit Model Log}
\usage{
compare_and_save_best_model(
  survey,
  model_comparison,
  name_fitting_res,
  mod_0,
  mod_1,
  mod_2
)
}
\arguments{
\item{data}{dat}
}
\value{
res
}
\description{
Función que ajusta el modelo logarítmico a los datos
Function that fits the logarithmic model to the data
}
