% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{dir_results}
\alias{dir_results}
\title{Dir Results}
\usage{
dir_results(name_dir)
}
\value{
dir
}
\description{
Función guarda los resultados en una dirección
Function saves the results in a place
}
