% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{extract_and_save}
\alias{extract_and_save}
\title{Extract and Save}
\usage{
extract_and_save(res_file_1, best_model_1, name_file, survey, real_yexpo)
}
\arguments{
\item{res}{res_file_3}

\item{best_model}{best_model_3}
}
\value{
extracted and saved results
}
\description{
Función que extrae y guarda resultados
Function that extracts and saves results
}
