% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{extract_model_summary}
\alias{extract_model_summary}
\title{Extract Model Summary}
\usage{
extract_model_summary(model_object)
}
\arguments{
\item{model_object}{what the run model function returns}

\item{model_data}{refers to data of the model}
}
\value{
summary of the models
}
\description{
Function that summarizes the models
}
\examples{
extract_model_summary (model_object)
}
