% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{extract_seroprev_model_summary}
\alias{extract_seroprev_model_summary}
\title{Extract Model Summary}
\usage{
extract_seroprev_model_summary(model_object)
}
\arguments{
\item{model_object}{\code{model_object}. An object containing relevant information about the implementation of the model. Refer to \link{fit_seroprev_model} for further details.}
}
\value{
\code{model_summary}. Object with a summary of \code{model_object} containing the following:
\tabular{ll}{
\code{seroprev_model_name} \tab Name of the selected model. For further details refer to \link{save_or_load_model}. \cr \tab \cr
\code{data_set} \tab Seroprevalence survey label.\cr \tab \cr
\code{country} \tab Name of the country were the survey was conducted in. \cr \tab \cr
\code{year} \tab Year in which the survey was conducted. \cr \tab \cr
\code{test} \tab Type of test of the survey. \cr \tab \cr
\code{antibody} \tab Antibody \cr \tab \cr
\code{n_sample} \tab Total number of samples in the survey. \cr \tab \cr
\code{n_agec} \tab Number of age groups considered. \cr \tab \cr
\code{n_iter} \tab Number of interations for eah chain including the warmup. \cr \tab \cr
\code{elpd} \tab elpd \cr \tab \cr
\code{se} \tab se \cr \tab \cr
\code{converged} \tab convergence \cr \tab \cr
}
}
\description{
Function to generate a summary of a model.
}
\examples{
\dontrun{
seroprev_data <- prepare_seroprev_data(mydata)
model_object <- run_seroprev_model(seroprev_data = seroprev_data,
                          seroprev_model_name = "constant_foi_bi")
extract_seroprev_model_summary (model_object)
}
}
