% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{extract_summary_mod}
\alias{extract_summary_mod}
\title{Extract Summary Mod}
\usage{
extract_summary_mod(res, dat)
}
\arguments{
\item{data}{data}
}
\value{
summary of the models
}
\description{
Función que hace un resumen de los modelos
Function that summarizes the models
}
