% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{extract_summary_model}
\alias{extract_summary_model}
\title{Extract Summary Model}
\usage{
extract_summary_model(model_object)
}
\arguments{
\item{model_data}{model_data}
}
\value{
summary of the models
}
\description{
Función que hace un resumen de los modelos
Function that summarizes the models
}
