% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_force.R
\name{fit_model}
\alias{fit_model}
\title{Fit Model}
\usage{
fit_model(
  model,
  dat,
  m_name,
  n_iters = 3000,
  n_thin = 2,
  delta = 0.9,
  mtreed = 10,
  Decades = 0
)
}
\arguments{
\item{data}{dat}
}
\value{
res
}
\description{
Función que ajusta el modelo a los datos
Function that fits the model to the data
}
