% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_model_log}
\alias{fit_model_log}
\title{Fit Model Log}
\usage{
fit_model_log(
  model,
  model_data,
  model_name,
  n_iters = 3000,
  n_thin = 2,
  delta = 0.9,
  m_treed = 10,
  decades = 0
)
}
\arguments{
\item{model}{refers to model selected}

\item{model_data}{model_data}

\item{model_name}{name of the model selected}

\item{n_iters}{number of iterations. This value comes by default but it can be changed}

\item{delta}{This value comes by default but it can be changed}

\item{m_treed}{This value comes by default but it can be changed}

\item{decades}{The decades covered by the survey data}

\item{n_thinThis}{value comes by default but it can be changed}
}
\value{
model_object
}
\description{
Function that fits the logarithmic model to the data
}
