% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_seromodel_optimization}
\alias{fit_seromodel_optimization}
\title{Fit model to seroprevalence survey using optimization
Returns a single best fit parameter vector, rather than
samples from the posterior (which is what \code{fit_seromodel} does)}
\usage{
fit_seromodel_optimization(
  serodata,
  foi_model = c("constant", "tv_normal_log", "tv_normal"),
  foi_location = 0,
  foi_scale = 1,
  chunks = NULL,
  chunk_size = 1,
  iter = 2000,
  seed = 12345,
  ...
)
}
\arguments{
\item{serodata}{A data frame containing the data from a seroprevalence
survey. This data frame must contain at least the following columns:
\describe{
\item{\code{total}}{Number of samples for each age group}
\item{\code{counts}}{Number of positive samples for each age group}
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{age_mean_f}}{Floor value of the average between age_min and age_max}
\item{\code{sample_size}}{The size of the sample}
\item{\code{birth_year}}{The year in which the individuals of each age group
were born}
}
The last six columns can be added to \code{serodata} by means of the function
\code{\link[=prepare_serodata]{prepare_serodata()}}.}

\item{foi_model}{Name of the selected model. Current version provides three
options:
\describe{
\item{\code{"constant"}}{Runs a constant model}
\item{\code{"tv_normal"}}{Runs a normal model}
\item{\code{"tv_normal_log"}}{Runs a normal logarithmic model}
}}

\item{foi_location}{Location parameter of the force-of-infection distribution
of the selected model. Depending on \code{foi_model}, the meaning may vary.}

\item{foi_scale}{Scale parameter of the force-of-infection distribution
of the selected model. Depending on \code{foi_model}, the meaning may vary.}

\item{chunks}{Numeric list specifying the chunk structure of the time
interval from the birth year of the oldest age cohort
\code{min(serodata$age_mean_f)} to the time when the serosurvey was conducted
\code{t_sur}. If \code{NULL}, the time interval is divided in chunks of size
\code{chunk_size}.}

\item{chunk_size}{Size of the chunks to be used in case that the chunk
structure \code{chunks} is not specified in \link{fit_seromodel}.
Default is 1, meaning that one force of infection value is to be estimated
for every year in the time interval spanned by the serosurvey.
If the length of the time interval is not exactly divisible by \code{chunk_size},
the remainder years are included in the last chunk.}

\item{iter}{maximum possible number of iterations for optimizer to run}

\item{seed}{For further details refer to the \code{seed} parameter in
\link[rstan:stanmodel-method-sampling]{sampling}.}

\item{...}{Additional parameters for \link[rstan:stanmodel-method-sampling]{sampling}.}
}
\value{
\code{seromodel_optimization}. List of best fit optimized parameter value
}
\description{
Fit model to seroprevalence survey using optimization
Returns a single best fit parameter vector, rather than
samples from the posterior (which is what \code{fit_seromodel} does)
}
