% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_combined_plots}
\alias{generate_combined_plots}
\title{Generate Combined Plots}
\usage{
generate_combined_plots(
  res,
  dat,
  lambda_sim = NA,
  max_lambda = NA,
  size_text = 25
)
}
\arguments{
\item{data}{Data}

\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
The combined plots
}
\description{
Función que genera la gráfica combinada
Function that generates the combined graph
}
