% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_prev_expanded_plot}
\alias{generate_prev_expanded_plot}
\title{Generate Previous Expanded Plot}
\usage{
generate_prev_expanded_plot(
  data,
  res,
  xlabel = "Age",
  ylabel = "Sero-positivity"
)
}
\arguments{
\item{data}{Data}

\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
The previous expanded graphic
}
\description{
Función que genera la gráfica previa expandida
Function that generates the previous expanded graphic
}
\examples{
generate_prev_expanded_plot(data, res, xlabel = "Age", ylabel = "Sero-positivity")
}
