% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_serosurvey.R
\name{generate_random_sample_sizes}
\alias{generate_random_sample_sizes}
\title{Generate random sample sizes for each age group.}
\usage{
generate_random_sample_sizes(survey_df_long)
}
\arguments{
\item{survey_df_long}{A dataframe with columns 'age', 'group' and
'overall_sample_size'.}
}
\value{
A dataframe with random sample sizes generated for each age based on
the overall sample size.
}
\description{
This function generates random sample sizes for each age group based on the
overall sample size and the distribution of individuals across age groups.
It uses multinomial sampling to allocate the total sample size to each age
group proportionally.
}
