% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{get_age_group}
\alias{get_age_group}
\title{Construct age-group variable from age column}
\usage{
get_age_group(age, step)
}
\arguments{
\item{age}{vector containing age information}

\item{step}{step used to split the age interval}
}
\value{
age_group factor variable grouping \code{age} by the age intervals
specified by \code{min(age)}, \code{max(age)}.
}
\description{
Simplified version of \link[vaccineff:get_age_group]{get_age_group}.
This function splits an age interval from age_min to age_max by
steps of length \code{step}.
\code{age_min} and \code{age_max} are calculated from \code{age}.
In cases that \code{age_max\%\%(step+1)!=0}, the last age interval is
truncated and will have a different length than the others.
}
