% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seromodel.R
\name{get_age_intervals}
\alias{get_age_intervals}
\title{Construct age-group variable from age column}
\usage{
get_age_intervals(serosurvey, step)
}
\arguments{
\item{serosurvey}{\describe{
\item{\code{survey_year}}{Year in which the survey took place
(only needed to plot time models)}
\item{\code{age_min}}{Floor value of the average between age_min and age_max}
\item{\code{age_max}}{The size of the sample}
\item{\code{sample_size}}{Number of samples for each age group}
\item{\code{n_seropositive}}{Number of positive samples for each age group}
}}

\item{step}{step used to split the age interval}
}
\value{
Serosurvey with addition factor variable grouping \code{age_intervals}.
The interval is taken as closed to the right and to the left.
}
\description{
Generates age intervals of length step in the interval spanned by
\code{age_min} and \code{age_max} in a serosurvey.
In cases where \code{max(age_max)\%\%(step+1)!=0}, the last age interval is
truncated and will have a different length than the others.
}
