% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_chunk_structure}
\alias{get_chunk_structure}
\title{Generate list containing the chunk structure to be used in the retrospective
estimation of the force of infection.}
\usage{
get_chunk_structure(serodata, chunk_size)
}
\arguments{
\item{serodata}{A data frame containing the data from a seroprevalence
survey. This data frame must contain at least the following columns:
\describe{
\item{\code{total}}{Number of samples for each age group}
\item{\code{counts}}{Number of positive samples for each age group}
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{age_mean_f}}{Floor value of the average between age_min and age_max}
\item{\code{sample_size}}{The size of the sample}
\item{\code{birth_year}}{The year in which the individuals of each age group
were born}
}
The last six columns can be added to \code{serodata} by means of the function
\code{\link[=prepare_serodata]{prepare_serodata()}}.}

\item{chunk_size}{Size of the chunks to be used in case that the chunk
structure \code{chunks} is not specified in \link{fit_seromodel}.
Default is 1, meaning that one force of infection value is to be estimated
for every year in the time interval spanned by the serosurvey.
If the length of the time interval is not exactly divisible by \code{chunk_size},
the remainder years are included in the last chunk.}
}
\description{
This function generates a numeric list specifying the chunk structure of the
time interval spanning from the year of birth of the oldest age cohort up to
the time when the serosurvey was conducted.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(serodata = chagas2012, alpha = 0.05)
cohort_ages <- get_cohort_ages(serodata = serodata)
}
