% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{get_comparison_table}
\alias{get_comparison_table}
\title{Get Model Table Comparison
Provides a table with statistics for comparison between models and selection}
\usage{
get_comparison_table(model_objects_list)
}
\arguments{
\item{model_objects_list}{model_objects to compare}
}
\value{
comparison table
}
\description{
Get Model Table Comparison
Provides a table with statistics for comparison between models and selection
}
\examples{
\dontrun{
data_test <- prepare_seroprev_data(serodata)
model_0 <- run_seroprev_model(seroprev_data = data_test,
                    seroprev_model_name = "constant_foi_bi",
                    n_iters = 1000)

model_1 <- run_seroprev_model(seroprev_data = data_test,
                    seroprev_model_name = "continuous_foi_normal_bi",
                    n_iters = 1000)

model_2 <- run_seroprev_model(seroprev_data = data_test,
                    seroprev_model_name = "continuous_foi_normal_log",
                    n_iters = 1000)
comp_table <- get_comparison_table(model_objects_list = c(m0 = model_0,
                                                          m1 = model_1,
                                                          m2 = model_2))
}
}
