% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{get_exposure_matrix}
\alias{get_exposure_matrix}
\title{Get Exposure Matrix}
\usage{
get_exposure_matrix(model_data, yexpo)
}
\arguments{
\item{model_data}{model_data}
}
\value{
exposure_output
}
\description{
Función que obtiene la matriz de exposición
Function that gets the exposure matrix
}
