% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stan_data.R
\name{get_foi_index}
\alias{get_foi_index}
\title{Generates force-of-infection indexes for heterogeneous age groups}
\usage{
get_foi_index(serosurvey, group_size)
}
\arguments{
\item{serosurvey}{\describe{
\item{\code{survey_year}}{Year in which the survey took place}
\item{\code{age_min}}{Floor value of the average between age_min and age_max}
\item{\code{age_max}}{The size of the sample}
\item{\code{sample_size}}{Number of samples for each age group}
\item{\code{n_seropositive}}{Number of positive samples for each age group}
}}

\item{group_size}{Age groups size}
}
\value{
Integer vector with the indexes numerating each year/age
(depending on the model).
}
\description{
The max value of the force-of-infection indexes correspond to
the number of foi values to be estimated when sampling.
}
