% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_model_comparison_plot}
\alias{get_model_comparison_plot}
\title{Get Model Comparison Plot}
\usage{
get_model_comparison_plot(res_comp)
}
\arguments{
\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
The model comparison graph
}
\description{
Función que obtiene la gráfica de comparación de modelos
Function that obtains the model comparison plot
}
