% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{get_posterior_summary}
\alias{get_posterior_summary}
\title{Get Posterior Summary}
\usage{
get_posterior_summary(model_objects_chain)
}
\arguments{
\item{model_objects}{model_objects_chain}
}
\value{
model_objects
}
\description{
Función que consigue un resumen posterior
Function that gets a post summary
}
