% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_force.R
\name{get_prev_expanded}
\alias{get_prev_expanded}
\title{Get Prevalence Expanded}
\usage{
get_prev_expanded(foi, dat)
}
\arguments{
\item{data}{dat}
}
\value{
prev_final
}
\description{
Función que obtiene la prevalencia expandida
Function that obtains the expanded prevalence
}
