% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{get_probability_exact_av}
\alias{get_probability_exact_av}
\title{Computes the probability of being seropositive for age-varying
FoI including seroreversion}
\usage{
get_probability_exact_av(exposure_age, foi, mu = 0)
}
\arguments{
\item{exposure_age}{Integer corresponding to the age of the exposed cohort}

\item{foi}{Numeric atomic vector corresponding to the age-varying FoI to
simulate from}

\item{mu}{Seroreversion rate}
}
\value{
probability of being seropositive for age-varying FoI
including seroreversion
}
\description{
Computes the probability of being seropositive for age-varying
FoI including seroreversion
}
