% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_force.R
\name{get_residuals}
\alias{get_residuals}
\title{Get Residuals}
\usage{
get_residuals(fit, dat)
}
\arguments{
\item{data}{dat}
}
\value{
merged_prev
}
\description{
Función que obtiene los residuos
Function that gets the residuals
}
