% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_serosurvey.R
\name{get_seroprev_counts_by_bin}
\alias{get_seroprev_counts_by_bin}
\title{Generate seropositivity counts by bin given the probability and sample size
per age group bin}
\usage{
get_seroprev_counts_by_bin(
  prob_seroprev_by_age,
  sample_size_by_age_random,
  survey_features
)
}
\arguments{
\item{prob_seroprev_by_age}{Probability of seropositivity by age}

\item{sample_size_by_age_random}{Random sample size by age}

\item{survey_features}{A dataframe containing information about the binned
age groups and sample sizes for each.
It should contain columns: \link{'age_min', 'age_max', 'n_sample'}.}
}
\description{
Generate seropositivity counts by bin given the probability and sample size
per age group bin
}
\keyword{internal}
