% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{get_sim_n_seropositive}
\alias{get_sim_n_seropositive}
\title{Function that generates a sample of counts of seropositive individuals by sampling from a binomial distribution}
\usage{
get_sim_n_seropositive(sim_data, foi, sample_size_by_age, seed = 1234)
}
\arguments{
\item{sim_data}{A dataframe object containing the following columns:
\tabular{ll}{
\code{tsur} \tab Year of the survey\cr \tab \cr
\code{age_mean_f} \tab Age \cr \tab \cr
\code{birth_year} \tab Years in which the subjects were borned
according to the age group marker \code{age_mean_f}\cr \tab \cr
}}

\item{foi}{Numeric atomic vector corresponding to the desired Force-of-Infection ordered from past to present}

\item{sample_size_by_age}{Sample size for each age group: either a single integer indicating that the sample size is 
the same for all ages or a vector of sample sizes the same length as
This corresponds to the number of trials \code{size} in \link[stats]{rbinom}.}

\item{seed}{The seed for random number generation.}
}
\value{
A dataframe containing the following columns:
\tabular{ll}{
\code{age} \tab Age by the time of the survey \cr \tab \cr
\code{n_seropositive} \tab Number of positive cases sampled according to the provided FoI \cr \tab \cr
}
simulated list of counts following a binomial distribution in accordance with a given
force of infection and age class sizes.
}
\description{
Function that generates a sample of counts of seropositive individuals by sampling from a binomial distribution
}
\examples{
\dontrun{
foi <- rep(0.02, 50)
sim_data <- generate_sim_data(foi = foi,
                              sample_size_by_age = 5,
                              tsur = 2050,
                              birth_year_min = 2000,
                              survey_label = 'foi_sim')
sim_n_seropositive <- get_sim_n_seropositive(sim_data = sim_data,
                                             foi = foi)
}
}
