% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{get_sim_prob}
\alias{get_sim_prob}
\title{Function that generates the probabilities of being previously exposed to a pathogen.}
\usage{
get_sim_prob(sim_data, foi, seed = 1234)
}
\arguments{
\item{sim_data}{A dataframe object containing the following columns:
\tabular{ll}{
\code{birth_year} \tab List of years in which the subjects were borned \cr \tab \cr
\code{tsur} \tab Year of the survey\cr \tab \cr
\code{country} \tab Default to 'none'.\cr \tab \cr
\code{survey} \tab Survey label \cr \tab \cr
\code{age_mean_f} \tab Age \cr \tab \cr
}}

\item{foi}{Numeric atomic vector corresponding to the desired Force-of-Infection.}

\item{seed}{The seed for random number generation.}
}
\value{
A simulated list containing a seropositivity distribution by age for given simulated
dataset and desired foi trend.
}
\description{
Function that generates the probabilities of being previously exposed to a pathogen.
}
\examples{
\dontrun{

}
}
