% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{get_table_rhats}
\alias{get_table_rhats}
\title{Get Table Rhats}
\usage{
get_table_rhats(seromodel_object)
}
\arguments{
\item{seromodel_object}{seromodel_object}
}
\value{
rhats table
}
\description{
Function that makes the rhats table
}
\examples{
\dontrun{
serodata <- prepare_serodata(serodata = serodata, alpha = 0.05)
seromodel_object <- run_seromodel(
serodata = serodata, seromodel_name = "constant_foi_bi")
get_table_rhats (seromodel_object)
}
}
