% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{make_thin_chain}
\alias{make_thin_chain}
\title{Make Thin Chain}
\usage{
make_thin_chain(model_objects_chain, thin = 10)
}
\arguments{
\item{thin}{by default the value 10 is taken but it can be changed}
}
\value{
model_objects_chain
}
\description{
Function that does the thinning of the number of iterations
}
