% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{make_yexpo}
\alias{make_yexpo}
\title{Make yexpo}
\usage{
make_yexpo(model_data)
}
\arguments{
\item{model_data}{refers to the model data that has been selected}
}
\value{
yexpo
}
\description{
Function that generates Yexpo
}
