% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_serosurvey.R
\name{multinomial_sampling_group}
\alias{multinomial_sampling_group}
\title{Generate random sample sizes using multinomial sampling.}
\usage{
multinomial_sampling_group(sample_size, n_ages)
}
\arguments{
\item{sample_size}{The total sample size to be distributed among age groups.}

\item{n_ages}{The number of age groups.}
}
\value{
A vector containing random sample sizes for each age group.
}
\description{
This function generates random sample sizes for each age group using
multinomial sampling. It takes the total sample size and the number of age
groups as input and returns a vector of sample sizes for each age group.
}
