% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{Seroprevalence data on serofoi}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
mydata
}
\description{
Data from a serological surveys
}
\examples{
mydata
}
\keyword{datasets}
