% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_force.R
\name{obtain_prevalence_extended}
\alias{obtain_prevalence_extended}
\title{Obtain Prevalence Extended}
\usage{
obtain_prevalence_extended(dato, exposure, ly, nbreaks, lambdaYexpo)
}
\arguments{
\item{data}{dat0}
}
\value{
new_PPP
}
\description{
Función que obtiene la prevalencia extendida
Function that obtains the extended prevalence
}
