% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_foi}
\alias{plot_foi}
\title{Generate force-of-infection plot corresponding to the
specified fitted serological model}
\usage{
plot_foi(
  seromodel_object,
  serodata,
  max_lambda = NA,
  size_text = 25,
  foi = NULL
)
}
\arguments{
\item{seromodel_object}{Stanfit object containing the results of fitting a
model by means of \link{fit_seromodel}.}

\item{serodata}{A data frame containing the data from a seroprevalence
survey. This data frame must contain at least the following columns:
\describe{
\item{\code{total}}{Number of samples for each age group}
\item{\code{counts}}{Number of positive samples for each age group}
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{age_mean_f}}{Floor value of the average between age_min and age_max}
\item{\code{sample_size}}{The size of the sample}
\item{\code{birth_year}}{The year in which the individuals of each age group
were born}
}
The last six columns can be added to \code{serodata} by means of the function
\code{\link[=prepare_serodata]{prepare_serodata()}}.}

\item{max_lambda}{Upper \code{ylim}for force-of-infection plot}

\item{size_text}{Text size use in the theme of the graph returned by the
function.}

\item{foi}{Data frame with the Force-of-infection trend to be plotted
alongside the estimated force-of-infection. Typically this corresponds to
the force-of-infection used to simulate the serosurvey used to model.}
}
\value{
A ggplot2 object containing the force-of-infection vs time including
the corresponding confidence interval.
}
\description{
This function generates a force-of-infection plot from the results obtained
by fitting a serological model. This includes the corresponding binomial
confidence interval. The x axis corresponds to the decades covered by the
survey the y axis to the force-of-infection.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_object <- fit_seromodel(
  serodata = serodata,
  foi_model = "constant",
  iter = 1000
)
cohort_ages <- get_cohort_ages(serodata)
plot_foi(
  seromodel_object = seromodel_object,
  cohort_ages = cohort_ages,
  size_text = 15
)
}
