% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seromodel.R
\name{plot_foi_estimates}
\alias{plot_foi_estimates}
\title{Plots force-of-infection central estimates}
\usage{
plot_foi_estimates(
  seromodel,
  serosurvey,
  alpha = 0.05,
  foi_df = NULL,
  foi_max = NULL,
  size_text = 11
)
}
\arguments{
\item{seromodel}{stan_fit object obtained from sampling a model
with \link{fit_seromode}}

\item{serosurvey}{\describe{
\item{\code{survey_year}}{Year in which the survey took place}
\item{\code{age_min}}{Floor value of the average between age_min and age_max}
\item{\code{age_max}}{The size of the sample}
\item{\code{sample_size}}{Number of samples for each age group}
\item{\code{n_seropositive}}{Number of positive samples for each age group}
}}

\item{alpha}{1 - alpha indicates the credibility level to be used}

\item{foi_df}{Dataframe with columns
\describe{
\item{\code{year}/\code{age}}{Year/Age (depending on the model)}
\item{\code{foi}}{Force-of-infection values by year/age}
}}

\item{foi_max}{Max force-of-infection value for plotting}

\item{size_text}{Size of text for plotting (\code{base_size} in
\link[ggplot2:ggtheme]{ggplot2})}
}
\value{
ggplot object with estimated force-of-infection
}
\description{
Plots force-of-infection central estimates
}
