% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_seroprev_model}
\alias{plot_seroprev_model}
\title{Generate a vertical arrange of plots summarizing the results of the model implementation}
\usage{
plot_seroprev_model(
  model_object,
  lambda_sim = NA,
  max_lambda = NA,
  size_text = 25
)
}
\arguments{
\item{model_object}{Object containing the results of fitting a model by means of \link{run_seroprev_model}.}

\item{size_text}{Text size use in the theme of the graph returned by the function.}
}
\value{
A ggplot object with a vertical arrange containing the seropositivity, force of infection, and convergence plots.
}
\description{
Function that generates the combined plots summarizing the results of the model implementation
}
\examples{
\dontrun{
data_test <- prepare_seroprev_data(serodata)
model_object <- run_seroprev_model(
 seroprev_data = data_test,
 seroprev_model_name = "constant_foi_bi",
 n_iters = 1000
)
plot_seroprev_model(model_object, size_text = 15)
}
}
