% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_seroprev_models_grid}
\alias{plot_seroprev_models_grid}
\title{Plot Several Models}
\usage{
plot_seroprev_models_grid(..., n_row = NULL, n_col = NULL)
}
\arguments{
\item{n_row}{Number of rows of the plot array.}

\item{n_col}{Number of columns of the plot array.}

\item{models_list}{A list containing an arbitrary number of model objects.
This objects can be obtained by means of \link{run_model}.}
}
\value{
A ggplot object containing an array with the plots in \code{models_list}.
}
\description{
Function that generates a plot showing the results for several models.
}
\examples{
\dontrun{
 data_test <- prepare_data(mydata)
 model_object_constant <- run_model(model_data = data_test,
                                    model_name = "constant_foi_bi")
 model_object_normalbi <- run_model(model_data = data_test,
                                    model_name = "continuous_foi_normal_bi")
 model_object_normallog <- run_model(model_data = data_test,
                                     model_name = "continuous_foi_normal_log")
 models_list <- c(model_object_constant, model_object_normalbi, model_object_normallog)
 plot_models_list(models_list, n_row = 1, n_col = 3)
}
}
