% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seromodel.R
\name{plot_seroprevalence_estimates}
\alias{plot_seroprevalence_estimates}
\title{Plot seroprevalence estimates on top of the serosurvey}
\usage{
plot_seroprevalence_estimates(seromodel, serosurvey, alpha = 0.05, ...)
}
\arguments{
\item{seromodel}{stan_fit object obtained from sampling a model
with \link{fit_seromode}}

\item{serosurvey}{\describe{
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{age_min}}{Floor value of the average between age_min and age_max}
\item{\code{age_max}}{The size of the sample}
\item{\code{sample_size}}{Number of samples for each age group}
\item{\code{n_seropositive}}{Number of positive samples for each age group}
}}

\item{alpha}{1 - alpha indicates the credibility level to be used}
}
\value{
ggplot object with seroprevalence estimates and serisurvey plots
}
\description{
Plot seroprevalence estimates on top of the serosurvey
}
