% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seromodel.R
\name{plot_serosurvey}
\alias{plot_serosurvey}
\title{Plots seroprevalence from the given serosurvey}
\usage{
plot_serosurvey(serosurvey, size_text = 11)
}
\arguments{
\item{serosurvey}{\describe{
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{age_min}}{Floor value of the average between age_min and age_max}
\item{\code{age_max}}{The size of the sample}
\item{\code{sample_size}}{Number of samples for each age group}
\item{\code{n_seropositive}}{Number of positive samples for each age group}
}}

\item{size_text}{Size of text for plotting (\code{base_size} in
\link[ggplot2:ggtheme]{ggplot2})}
}
\value{
ggplot object with seroprevalence plot
}
\description{
Plots seroprevalence from the given serosurvey
}
