% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seromodel.R
\name{plot_summary}
\alias{plot_summary}
\title{Plots model summary}
\usage{
plot_summary(
  seromodel,
  serosurvey,
  elpd_digits = 1,
  foi_digits = 2,
  seroreversion_digits = 2,
  rhat_digits = 2
)
}
\arguments{
\item{seromodel}{stan_fit object obtained from sampling a model
with \link{fit_seromode}}

\item{serosurvey}{\describe{
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{age_min}}{Floor value of the average between age_min and age_max}
\item{\code{age_max}}{The size of the sample}
\item{\code{sample_size}}{Number of samples for each age group}
\item{\code{n_seropositive}}{Number of positive samples for each age group}
}}

\item{elpd_digits}{Number of elpd digits to show}

\item{foi_digits}{Number of foi digits to show}

\item{seroreversion_digits}{Number of seroreversion rate digits to show}
}
\value{
ggplot object with a summary of the specified model
}
\description{
Plots model summary
}
