% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data}
\usage{
prepare_data(model_data, alpha = 0.05)
}
\arguments{
\item{model_data}{dataset to be processed}

\item{alpha}{probability of a type I error (Hmisc::binconf)}
}
\value{
model_data with additional columns necessary for the analysis
}
\description{
Function that prepares the data for modelling
}
