% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{prepare_serodata}
\alias{prepare_serodata}
\title{Prepare data from a serological survey for modelling}
\usage{
prepare_serodata(serodata = serodata, alpha = 0.05)
}
\arguments{
\item{serodata}{A data frame containing the data from a serological survey.
This data frame must contain the following columns:
\describe{
\item{\code{survey}}{survey Label of the current survey}
\item{\code{total}}{Number of samples for each age group}
\item{\code{counts}}{Number of positive samples for each age group}
\item{\code{age_min}}{age_min}
\item{\code{age_max}}{age_max}
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{country}}{The country where the survey took place}
\item{\code{test}}{The type of test taken}
\item{\code{antibody}}{antibody}
}
Alternatively to \code{age_min} and \code{age_max}, the dataset could already include
the age group marker \code{age_mean_f}, representing the middle point between
\code{age_min} and \code{age_max}. If \code{afe_mean_f} is missing, it will be generated
by the function.}

\item{alpha}{probability of a type I error. For further details refer to
\link[Hmisc:binconf]{binconf}.}
}
\value{
serodata with additional columns necessary for the analysis. These
columns are:
\describe{
\item{\code{age_mean_f}}{Floor value of the average between age_min and age_max
for the age groups delimited by \code{age_min} and \code{age_max}}
\item{\code{sample_size}}{The size of the sample}
\item{\code{birth_year}}{Years in which the subject was born according to the
age group marker \code{age_mean_f}}
\item{\code{prev_obs}}{Observed prevalence}
\item{\code{prev_obs_lower}}{Lower limit of the confidence interval for the
observed prevalence}
\item{\code{prev_obs_upper}}{Upper limit of the confidence interval for the
observed prevalence}
}
}
\description{
This function adds the necessary additional variables to the given dataset
\code{serodata} corresponding to a serological survey.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(chagas2012)
}
