% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{preprare_seroprev_data}
\alias{preprare_seroprev_data}
\title{Prepare data}
\usage{
preprare_seroprev_data(seroprev_data, alpha = 0.05)
}
\arguments{
\item{seroprev_data}{A data frame containing the data from a seroprevalence survey.
This data frame must contain the following columns:
\tabular{ll}{
\code{survey} \tab survey Label of the current survey \cr \tab \cr
\code{total} \tab Number of samples for each age group\cr \tab \cr
\code{counts} \tab Number of positive samples for each age group\cr \tab \cr
\code{age_min} \tab age_min \cr \tab \cr
\code{age_max} \tab age_max \cr \tab \cr
\code{year_init} \tab year_init \cr \tab \cr
\code{year_end} \tab year_end \cr \tab \cr
\code{tsur} \tab Year in which the survey took place \cr \tab \cr
\code{country} \tab The country where the survey took place \cr \tab \cr
\code{test} \tab The type of test taken \cr \tab \cr
\code{antibody} \tab antibody \cr \tab \cr
}}

\item{alpha}{probability of a type I error. For further details refer to \link{Hmisc::binconf}.}
}
\value{
seroprev_data with additional columns necessary for the analysis. These columns are:
\tabular{ll}{
\code{age_mean_f} \tab Floor value of the average between age_min and age_max \cr \tab \cr
\code{sample_size} \tab The size of the sample \cr \tab \cr
\code{birth_year} \tab The year in which the individuals of each age group were bornt \cr \tab \cr
\code{prev_obs} \tab Observed prevalence \cr \tab \cr
\code{prev_obs_lower} \tab Lower limit of the confidence interval for the observed prevalence \cr \tab \cr
\code{prev_obs_upper} \tab Upper limit of the confidence interval for the observed prevalence \cr \tab \cr
}
}
\description{
Function that prepares the data for modelling
}
\examples{
\dontrun{
data_test <- readRDS("data/data.RDS")
data_test <- preprare_seroprev_data(seroprev_data, alpha)
}
}
