% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_serosurvey.R
\name{prob_seroprev_age_time_by_age}
\alias{prob_seroprev_age_time_by_age}
\title{Generate probabilities of seropositivity by age based on an age-and-time
varying Force-of-Infection (FoI) model.}
\usage{
prob_seroprev_age_time_by_age(foi, seroreversion_rate)
}
\arguments{
\item{foi}{A dataframe containing the FoI values
for different ages. It should have three columns: 'year', 'age' and 'foi'.}

\item{seroreversion_rate}{A non-negative numeric value representing
the rate of seroreversion.}
}
\value{
A dataframe with columns 'age' and 'seropositivity'.
}
\description{
This function calculates the probabilities of seropositivity by age based on
an age-and-time-varying FoI model.
It takes into account the FoI and the rate of seroreversion.
}
