% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_serosurvey.R
\name{probability_seropositive_general_model_by_age}
\alias{probability_seropositive_general_model_by_age}
\title{Generate probabilities of seropositivity by age based on a general FOI model.}
\usage{
probability_seropositive_general_model_by_age(
  construct_A_fn,
  calculate_seropositivity_function,
  initial_conditions,
  max_age,
  ...
)
}
\arguments{
\item{construct_A_fn}{A function that constructs a matrix that defines the
multiplier term in the linear ODE system.}

\item{calculate_seropositivity_function}{A function which takes the state
vector and returns the seropositive fraction.}

\item{initial_conditions}{The initial state vector proportions for each
birth cohort.}

\item{max_age}{The maximum age to simulate seropositivity for.}
}
\value{
A dataframe with columns 'age' and 'seropositivity'.
}
\description{
This function calculates the probabilities of seropositivity by age based on
an abstract model of the serocatalytic system.
}
