% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{run_model}
\alias{run_model}
\title{Run model
Function that runs the specified model}
\usage{
run_model(
  model_data,
  model_name = "constant_foi_bi",
  n_iters = 1000,
  n_thin = 2,
  delta = 0.9,
  m_treed = 10,
  decades = 0
)
}
\arguments{
\item{model_data}{model_data}

\item{model_name}{name of the model selected}

\item{n_iters}{number of iterations. Each model has a value by default.}

\item{survey}{survey}

\item{model}{refers to model selected}
}
\value{
model_object of model
}
\description{
Run model
Function that runs the specified model
}
