% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{run_model_0}
\alias{run_model_0}
\title{Run Model 0}
\usage{
run_model_0(my_dir, survey, dat0, n_iters, n_warmup, model_0, name_model_0)
}
\arguments{
\item{data}{dat0}

\item{model}{model_0}

\item{name_model}{name_model_0}
}
\value{
results of model 0
}
\description{
Función que corre el modelo 0
Function that runs model 0
}
