% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_studies.R
\name{run_save_models}
\alias{run_save_models}
\title{Run Save Models}
\usage{
run_save_models(
  my_dir,
  suv,
  dat0,
  n_iters,
  n_warmup,
  Model0,
  NameModel0,
  Model1,
  NameModel1,
  Model2,
  NameModel2
)
}
\arguments{
\item{data}{dat0}

\item{model}{Model2}

\item{name_model}{NameModel2}
}
\value{
run save models
}
\description{
Función que corre los modelos guardados
Function that runs saved models
}
