% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{run_seroprev_model}
\alias{run_seroprev_model}
\title{Run the specified stan model for the force-of-infection}
\usage{
run_seroprev_model(
  seroprev_data,
  seroprev_model_name = "constant_foi_bi",
  n_iters = 1000,
  n_thin = 2,
  delta = 0.9,
  m_treed = 10,
  decades = 0
)
}
\arguments{
\item{seroprev_data}{A data frame containing the data from a seroprevalence survey.
This data frame must contain the following columns:
\tabular{ll}{
\code{survey} \tab survey Label of the current survey \cr \tab \cr
\code{total} \tab Number of samples for each age group\cr \tab \cr
\code{counts} \tab Number of positive samples for each age group\cr \tab \cr
\code{age_min} \tab age_min \cr \tab \cr
\code{age_max} \tab age_max \cr \tab \cr
\code{year_init} \tab year_init \cr \tab \cr
\code{year_end} \tab year_end \cr \tab \cr
\code{tsur} \tab Year in which the survey took place \cr \tab \cr
\code{country} \tab The country where the survey took place \cr \tab \cr
\code{test} \tab The type of test taken \cr \tab \cr
\code{antibody} \tab antibody \cr \tab \cr
\code{age_mean_f} \tab Floor value of the average between age_min and age_max \cr \tab \cr
\code{sample_size} \tab The size of the sample \cr \tab \cr
\code{birth_year} \tab The year in which the individuals of each age group were bornt \cr \tab \cr
\code{prev_obs} \tab Observed prevalence \cr \tab \cr
\code{prev_obs_lower} \tab Lower limit of the confidence interval for the observed prevalence \cr \tab \cr
\code{prev_obs_upper} \tab Upper limit of the confidence interval for the observed prevalence \cr \tab \cr
}
The last six colums can be added to \code{seroprev_data} by means of the function \code{\link{preprare_seroprev_data}}.}

\item{seroprev_model_name}{Name of the selected model. Current version provides three options:
\describe{
\item{\code{"constant_foi_bi"}}{Runs a constant model}
\item{\code{"continuous_foi_normal_bi"}}{Runs a normal model}
\item{\code{"continuous_foi_normal_log"}}{Runs a normal logarithmic model}
}}

\item{n_iters}{Number of interations for eah chain including the warmup. \code{iter} in \link[rstan]{sampling}.}

\item{n_thin}{Positive integer specifying the period for saving samples. \code{thin} in \link[rstan]{sampling}.}

\item{delta}{Real number between 0 and 1 that represents the target average acceptance probability.
Increasing the value of \code{delta} will result in a smaller step size and fewer divergences.
For further details refer to the \code{control} parameter in \link[rstan]{sampling} or \href{https://mc-stan.org/rstanarm/reference/adapt_delta.html}{here}.}

\item{m_treed}{Maximum tree depth for the binary tree used in the NUTS stan sampler. For further details refer to the \code{control} parameter in \link[rstan]{sampling}.}

\item{decades}{Number of decades covered by the survey data.}
}
\value{
\code{model_object}. An object containing relevant information about the implementation of the model. For further details refer to \link{fit_seroprev_model}.
}
\description{
Run the specified stan model for the force-of-infection
}
\examples{
\dontrun{
seroprev_data <- preprare_seroprev_data(mydata)
run_seroprev_model (seroprev_data,
           seroprev_model_name = "constant_foi_bi")
}
}
