% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_serosurvey.R
\name{sample_size_by_individual_age_random}
\alias{sample_size_by_individual_age_random}
\title{Generate random sample sizes for each individual age based on survey
features.}
\usage{
sample_size_by_individual_age_random(survey_features)
}
\arguments{
\item{survey_features}{A dataframe containing information about individuals'
age ranges and sample sizes.}
}
\value{
A dataframe with random sample sizes generated for each individual
age based on the provided survey features.
}
\description{
This function generates random sample sizes for each individual age based on
the provided survey features. It first creates age bins, assigns each
individual in the survey features to an age bin, calculates the overall
sample size by group, and then generates random sample sizes for each age
group. Finally, it returns a dataframe with the random sample sizes for each
individual age.
}
