% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{save_or_load_model}
\alias{save_or_load_model}
\title{Function used to determine whether the stan model corresponding to the specified serological model has been already compiled or not}
\usage{
save_or_load_model(foi_model = "constant")
}
\arguments{
\item{foi_model}{Name of the selected model. Current version provides three options:
\describe{
\item{\code{"constant"}}{Runs a constant model}
\item{\code{"tv_normal"}}{Runs a normal model}
\item{\code{"tv_normal_log"}}{Runs a normal logarithmic model}
}}
}
\value{
\code{model}. The rstan model object corresponding to the selected model.
}
\description{
This function determines whether the corresponding .RDS file of the selected model exists or not.
In case the .RDS file exists, it is read and returned; otherwise, the object model is created through the
\link[rstan]{stan_model} function, saved as an .RDS file and returned as the output of the function.
}
\examples{
\dontrun{
model <- save_or_load_model(foi_model="constant")
}

}
