% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{save_or_read_model}
\alias{save_or_read_model}
\title{Función que guarda el archivo .RDS del modelo
Function that saves the .RDS file of the model}
\usage{
save_or_read_model(model_name = "constant_foi_bi")
}
\arguments{
\item{model_name}{name of the model}
}
\description{
Función que guarda el archivo .RDS del modelo
Function that saves the .RDS file of the model
}
