% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serodata.R
\docType{data}
\name{serodata}
\alias{serodata}
\title{Seroprevalence data on serofoi}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
serodata
}
\description{
Data from a serological surveys
}
\examples{
serodata
}
\keyword{datasets}
