% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serodata_simD.R
\docType{data}
\name{serodata_simD}
\alias{serodata_simD}
\title{Seroprevalence data on serofoi}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
serodata_simD
}
\description{
Data from a serological surveys
}
\examples{
serodata_simD
}
\keyword{datasets}
