% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serofoi-package.R
\docType{package}
\name{serofoi-package}
\alias{serofoi-package}
\alias{serofoi}
\title{The 'serofoi' package.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimating the force of infection from time varying, age varying, or constant serocatalytic models from population based seroprevalence studies using a Bayesian framework, including data simulation functions enabling the generation of serological surveys based on this models. This tool also provides a flexible prior specification syntax for the force of infection and the seroreversion rate, as well as methods to assess model convergence and comparison criteria along with useful visualisation functions.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R
package version 2.26.22. https://mc-stan.org

#' @keywords internal
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/serofoi}
  \item \url{https://epiverse-trace.github.io/serofoi/}
  \item Report bugs at \url{https://github.com/epiverse-trace/serofoi/issues}
}

}
\author{
\strong{Maintainer}: Zulma M. Cucunubá \email{zulma.cucunuba@javeriana.edu.co} (\href{https://orcid.org/0000-0002-8165-3198}{ORCID})

Authors:
\itemize{
  \item Nicolás T. Domínguez \email{ex-ntorres@javeriana.edu.co} (\href{https://orcid.org/0009-0002-8484-1298}{ORCID})
  \item Ben Lambert
  \item Pierre Nouvellet
}

Other contributors:
\itemize{
  \item Geraldine Gómez (\href{https://orcid.org/0009-0007-8701-0568}{ORCID}) [contributor]
  \item Jaime A. Pavlich-Mariscal (\href{https://orcid.org/0000-0002-3892-6680}{ORCID}) [contributor]
  \item Hugo Gruson (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [contributor]
  \item David Santiago Quevedo (\href{https://orcid.org/0000-0003-1583-4262}{ORCID}) [contributor]
  \item Miguel Gámez [contributor]
  \item Sumali Bajaj [contributor]
  \item Everlyn Kamau [contributor]
  \item Richard Creswell [contributor]
  \item International Development Research Center (IDRC) [funder]
  \item Pontificia Universidad Javeriana [copyright holder]
}

}
