% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stan_data.R
\name{sf_cauchy}
\alias{sf_cauchy}
\title{Sets Cauchy distribution parameters for sampling}
\usage{
sf_cauchy(location = 0, scale = 1)
}
\arguments{
\item{scale}{Scale
of the normal distribution}

\item{sd}{Standard deviation of the normal distribution}
}
\value{
List with specified statistics and name of the model
}
\description{
Sets Cauchy distribution parameters for sampling
}
