% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stan_data.R
\name{sf_uniform}
\alias{sf_uniform}
\title{Sets uniform distribution parameters for sampling}
\usage{
sf_uniform(min = 0, max = 10)
}
\arguments{
\item{min}{Minimum value of the random variable of the uniform distribution}

\item{max}{Maximum value of the random variable of the uniform distribution}
}
\value{
List with specified statistics and name of the model
}
\description{
Sets uniform distribution parameters for sampling
}
\examples{
my_prior <- sf_uniform()
}
