% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata_large_epi.R
\docType{data}
\name{simdata_large_epi}
\alias{simdata_large_epi}
\title{Large epidemic simulated serosurvey}
\format{
An object of class \code{"cross"}; see \code{\link[qtl:read.cross]{qtl::read.cross()}}.
}
\usage{
simdata_large_epi
}
\description{
Simulated dataset describing a large epidemic between the years 2032 and 2035
with a constant force-of-infection with value 1.5.
The hypothetical serosurvey is conducted in year 2050 for 250 individuals
that are up to 50 years old.
}
\examples{
simdata_large_epi
}
\keyword{datasets}
