% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata_sw_dec.R
\docType{data}
\name{simdata_sw_dec}
\alias{simdata_sw_dec}
\title{Step-wise decreasing FoI simulated serosurvey}
\format{
An object of class \code{"cross"}; see \code{\link[qtl:read.cross]{qtl::read.cross()}}.
}
\usage{
simdata_sw_dec
}
\description{
Simulated dataset describing a situation where the Force-of-Infection
follows a step-wise decreasing tendency over a period of 50 years (2000-2050)
The hypothetical serosurvey is conducted in year 2050 for 250 individuals
that are up to 50 years old.
}
\examples{
simdata_sw_dec
}
\keyword{datasets}
