% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_serosurvey.R
\name{simulate_serosurvey_age}
\alias{simulate_serosurvey_age}
\title{Simulate serosurvey data based on an age-varying FOI model.}
\usage{
simulate_serosurvey_age(foi, survey_features, seroreversion_rate = 0)
}
\arguments{
\item{foi}{A dataframe containing the force of infection (FOI) values.
For time-varying models the columns should be \link{'year', 'foi'}.
For age-varying models the columns should be \link{'age', 'foi'}.
For age-and-time-varying models the columns should be \link{'age', 'time', 'foi'}.}

\item{survey_features}{A dataframe containing information about the binned
age groups and sample sizes for each.
It should contain columns: \link{'age_min', 'age_max', 'n_sample'}.}

\item{seroreversion_rate}{A non-negative value determining the rate of
seroreversion (per year). Default is 0.}
}
\value{
A dataframe with simulated serosurvey data, including age group
information, overall sample sizes, the number of seropositive individuals,
and other survey features.
}
\description{
This function generates binned serosurvey data based on an age-varying FOI
model, optionally including seroreversion. This function allows construction
of serosurveys with binned age groups, and it generates uncertainty in the
distribution of a sample size within an age bin through multinomial sampling.
}
\examples{
# specify FOIs for each year
foi_df <- data.frame(
  age = seq(1, 20, 1),
  foi = rnorm(20, 0.1, 0.01)
)
survey_features <- data.frame(
  age_min = c(1, 3, 15),
  age_max = c(2, 14, 20),
  n_sample = c(1000, 2000, 1500))
serosurvey <- simulate_serosurvey_age(
foi_df, survey_features)
}
