% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_serosurvey.R
\name{simulate_serosurvey_age_and_time_model}
\alias{simulate_serosurvey_age_and_time_model}
\title{Simulate serosurvey data based on an age-and-time-varying FOI model.}
\usage{
simulate_serosurvey_age_and_time_model(
  foi,
  survey_features,
  seroreversion_rate = 0
)
}
\arguments{
\item{foi}{A dataframe containing the force of infection (FOI) values for different ages.
It should have two columns: 'year', 'age' and 'foi'.}

\item{survey_features}{A dataframe containing information about the binned age groups and sample
sizes for each. It should contain columns: \link{'age_min', 'age_max', 'sample_size'}.}

\item{seroreversion_rate}{A non-negative value determining the rate of seroreversion (per year).
Default is 0.}
}
\value{
A dataframe with simulated serosurvey data, including age group information, overall
sample sizes, the number of seropositive individuals, and other survey features.
}
\description{
This function generates binned serosurvey data based on an age-and-time-varying FOI model,
optionally including seroreversion. This function allows construction of serosurveys
with binned age groups, and it generates uncertainty in the distribution of a sample size
within an age bin through multinomial sampling.
}
\examples{
# specify FOIs for each year
foi_df <- data.frame(
  year = seq(1990, 2009, 1),
  age = seq(1, 20, 1)
) \%>\%
mutate(foi = rnorm(20 * 20, 0.1, 0.01))
survey_features <- data.frame(
  age_min = c(1, 3, 15),
  age_max = c(2, 14, 20),
  sample_size = c(1000, 2000, 1500))
serosurvey <- simulate_serosurvey_age_and_time_model(
foi_df, survey_features)
}
