% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_seromodel.R
\name{summarise_loo_estimate}
\alias{summarise_loo_estimate}
\title{Extract specified loo estimate}
\usage{
summarise_loo_estimate(
  seromodel,
  par_loo_estimate = "elpd_loo",
  loo_estimate_digits = 2
)
}
\arguments{
\item{seromodel}{stan_fit object obtained from sampling a model
with \link{fit_seromode}}

\item{par_loo_estimate}{Name of the loo estimate to be extracted.}

\item{loo_estimate_digits}{Number of loo estimate digits}
}
\value{
Text summarising specified loo estimate
}
\description{
Extract specified loo estimate
}
