% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seromodel.R
\name{summarise_seromodel}
\alias{summarise_seromodel}
\title{Summarise specified model}
\usage{
summarise_seromodel(
  seromodel,
  serosurvey,
  alpha = 0.05,
  elpd_digits = 1,
  foi_digits = 2,
  seroreversion_digits = 2,
  rhat_digits = 2
)
}
\arguments{
\item{seromodel}{stan_fit object obtained from sampling a model
with \link{fit_seromode}}

\item{serosurvey}{\describe{
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{age_min}}{Floor value of the average between age_min and age_max}
\item{\code{age_max}}{The size of the sample}
\item{\code{sample_size}}{Number of samples for each age group}
\item{\code{n_seropositive}}{Number of positive samples for each age group}
}}

\item{alpha}{1 - alpha indicates the credibility level to be used}

\item{elpd_digits}{Number of elpd digits to show}

\item{foi_digits}{Number of foi digits to show}

\item{seroreversion_digits}{Number of seroreversion rate digits to show}
}
\value{
A list showing
\describe{
\item{\code{model_name}}{Name of the model}
\item{\code{elpd}}{elpd and its standard deviation}
\item{\code{foi}}{Estimated foi with credible interval (for 'constant' model)}
\item{\code{foi_rhat}}{foi rhat value (for 'constant' model)}
\item{\code{seroreversion_rate}}{Estimated seroreversion rate}
\item{\code{rhat_seroreversion_rate}}{Seroreversion rate rhat value}
}
}
\description{
Summarise specified model
}
