% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_serosurvey.R
\name{survey_by_individual_age}
\alias{survey_by_individual_age}
\title{Create a survey dataframe with per individual age information.}
\usage{
survey_by_individual_age(survey_features, age_df)
}
\arguments{
\item{survey_features}{A dataframe containing information about age groups
and sample sizes.}

\item{age_df}{A dataframe containing 'age' and 'group'.}
}
\value{
A dataframe with overall sample sizes calculated by joining
survey_features and age_df.
This dataframe has columns including 'age' and 'overall_sample_size'.
}
\description{
Create a survey dataframe with per individual age information.
}
