% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_foi_df}
\alias{validate_foi_df}
\title{Helper function to validate FOI structure for simulation}
\usage{
validate_foi_df(foi_df, cnames_additional)
}
\value{
None
}
\description{
Helper function to validate FOI structure for simulation
}
\keyword{internal}
