% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_foi_index}
\alias{validate_foi_index}
\title{Helper function to validate FoI index consistency}
\usage{
validate_foi_index(foi_index, serosurvey, model_type)
}
\value{
foi_index
}
\description{
Helper function to validate FoI index consistency
}
\keyword{internal}
