% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_plot_constant}
\alias{validate_plot_constant}
\title{Helper function to validate whether the current plot corresponds
to a constant model}
\usage{
validate_plot_constant(plot_constant, x_axis, model_name, error_msg_x_axis)
}
\value{
TRUE
}
\description{
Helper function to validate whether the current plot corresponds
to a constant model
}
\keyword{internal}
