% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_seroreversion_rate}
\alias{validate_seroreversion_rate}
\title{Helper function to validate seroreversion rate properties for simulation}
\usage{
validate_seroreversion_rate(seroreversion_rate)
}
\value{
None
}
\description{
Helper function to validate seroreversion rate properties for simulation
}
\keyword{internal}
