% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_serosurvey}
\alias{validate_serosurvey}
\title{Helper function to validate serosurvey structure}
\usage{
validate_serosurvey(serosurvey)
}
\description{
Helper function to validate serosurvey structure
}
\keyword{internal}
