% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_simulation_age}
\alias{validate_simulation_age}
\title{Helper function to validate consistency between the FOI and the survey
features for simulation}
\usage{
validate_simulation_age(survey_features, foi_df)
}
\value{
None
}
\description{
Helper function to validate consistency between the FOI and the survey
features for simulation
}
\keyword{internal}
