% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_simulation_age_time}
\alias{validate_simulation_age_time}
\title{Helper function to validate consistency between the FOI and the survey
features for simulation of age- and time-varying model}
\usage{
validate_simulation_age_time(survey_features, foi_df)
}
\value{
None
}
\description{
Helper function to validate consistency between the FOI and the survey
features for simulation of age- and time-varying model
}
\keyword{internal}
