% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_survey_features}
\alias{validate_survey_features}
\title{Helper function to validate serosurvey features for simulation}
\usage{
validate_survey_features(survey_features)
}
\value{
None
}
\description{
Helper function to validate serosurvey features for simulation
}
\keyword{internal}
