% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veev2012.R
\docType{data}
\name{veev2012}
\alias{veev2012}
\title{Venezuelan Equine Encephalitis Virus (VEEV) seroprevalence data in serofoi}
\format{
\subsection{\code{veev2012}}{

A \verb{<data.frame>} with 4 rows and 5 columns:
\describe{
\item{survey_year}{Year in which the serosurvey was conducted}
\item{n_sample}{Number of collected samples per age group}
\item{n_seropositive}{Number of positive samples per age group}
\item{age_min}{Age group minimal age}
\item{age_max}{Age group maximal age}
}
}
}
\usage{
data(veev2012)
}
\description{
Datasets that measure the seroprevalence of IgG antibodies against VEEV in
a rural village in Panamá in 2012. \link{carrera2020}
}
\examples{
data(veev2012)
}
\keyword{datasets}
