% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{vertical_plot_arrange_per_model}
\alias{vertical_plot_arrange_per_model}
\title{Get Vertical Plot Arrange per Model}
\usage{
vertical_plot_arrange_per_model(PPC)
}
\arguments{
\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
The vertical plot arrange per model
}
\description{
Función que genera el grafico en un arreglo vertical por modelo
Function that generates the graph in a vertical arrange per model
}
