% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chagas2012.R
\docType{data}
\name{chagas2012}
\alias{chagas2012}
\title{Seroprevalence data on serofoi}
\format{
An object of class \code{"cross"}; see \code{\link[qtl:read.cross]{qtl::read.cross()}}.
}
\usage{
chagas2012
}
\description{
Data from a serological surveys
}
\examples{
chagas2012
}
\keyword{datasets}
