% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chik2015.R
\docType{data}
\name{chik2015}
\alias{chik2015}
\title{Seroprevalence data on serofoi}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
chik2015
}
\description{
Data from a serological surveys
}
\examples{
chik2015
}
\keyword{datasets}
