% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{compare_and_save_best_model}
\alias{compare_and_save_best_model}
\title{Compare and Save Best Model}
\usage{
compare_and_save_best_model(
  survey,
  model_comparison,
  name_fitting_result,
  model_0
)
}
\arguments{
\item{mod}{model_2}
}
\value{
comparison and saving of the best model
}
\description{
Función que compara y salva el mejor modelo
Function that compares and saves the best model
}
