% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{extract_seromodel_summary}
\alias{extract_seromodel_summary}
\title{Method to extact a summary of the specified serological model object}
\usage{
extract_seromodel_summary(seromodel_object)
}
\arguments{
\item{seromodel_object}{\code{seromodel_object}. An object containing relevant information about the implementation of the model.
Refer to \link{fit_seromodel} for further details.}
}
\value{
\code{model_summary}. Object with a summary of \code{seromodel_object} containing the following:
\tabular{ll}{
\code{foi_model} \tab Name of the selected model. \cr \tab \cr
\code{data_set} \tab Seroprevalence survey label.\cr \tab \cr
\code{country} \tab Name of the country were the survey was conducted in. \cr \tab \cr
\code{year} \tab Year in which the survey was conducted. \cr \tab \cr
\code{test} \tab Type of test of the survey. \cr \tab \cr
\code{antibody} \tab Antibody \cr \tab \cr
\code{n_sample} \tab Total number of samples in the survey. \cr \tab \cr
\code{n_agec} \tab Number of age groups considered. \cr \tab \cr
\code{n_iter} \tab Number of interations for eah chain including the warmup. \cr \tab \cr
\code{elpd} \tab elpd \cr \tab \cr
\code{se} \tab se \cr \tab \cr
\code{converged} \tab convergence \cr \tab \cr
}
}
\description{
This method extracts a summary corresponding to a serological model object that contains information about the original serological
survey data used to fit the model, such as the year when the survey took place, the type of test taken and the corresponding antibody,
as well as information about the convergence of the model, like the expected log pointwise predictive density \code{elpd} and its
corresponding standar deviation.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_object <- run_seromodel(serodata = serodata,
                                  foi_model = "constant")
extract_seromodel_summary(seromodel_object)
}
