% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{extract_seromodel_summary}
\alias{extract_seromodel_summary}
\title{Function to extract a summary of the specified serological model object}
\usage{
extract_seromodel_summary(seromodel_object, serodata)
}
\arguments{
\item{seromodel_object}{Stanfit object containing the results of fitting a
model by means of \link{run_seromodel}.}

\item{serodata}{A data frame containing the data from a seroprevalence
survey. This data frame must contain the following columns:
\describe{
\item{\code{survey}}{survey Label of the current survey}
\item{\code{total}}{Number of samples for each age group}
\item{\code{counts}}{Number of positive samples for each age group}
\item{\code{age_min}}{age_min}
\item{\code{age_max}}{age_max}
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{country}}{The country where the survey took place}
\item{\code{test}}{The type of test taken}
\item{\code{antibody}}{antibody}
\item{\code{age_mean_f}}{Floor value of the average between age_min and age_max}
\item{\code{sample_size}}{The size of the sample}
\item{\code{birth_year}}{The year in which the individuals of each age group
were born}
\item{\code{prev_obs}}{Observed prevalence}
\item{\code{prev_obs_lower}}{Lower limit of the confidence interval for the
observed prevalence}
\item{\code{prev_obs_upper}}{Upper limit of the confidence interval for the
observed prevalence}
}
The last six columns can be added to \code{serodata} by means of the function
\code{\link[=prepare_serodata]{prepare_serodata()}}.}
}
\value{
\code{model_summary}. Object with a summary of \code{seromodel_object}
containing the following:
\describe{
\item{\code{foi_model}}{Name of the selected model.}
\item{\code{data_set}}{Seroprevalence survey label}
\item{\code{country}}{Name of the country were the survey was conducted in.}
\item{\code{year}}{Year in which the survey was conducted.}
\item{\code{test}}{Type of test of the survey.}
\item{\code{antibody}}{Antibody}
\item{\code{n_sample}}{Total number of samples in the survey.}
\item{\code{n_agec}}{Number of age groups considered.}
\item{\code{n_iter}}{Number of interactions for each chain including warmup.}
\item{\code{elpd}}{elpd}
\item{\code{se}}{se}
\item{\code{converged}}{convergence}
}
}
\description{
This function extracts a summary corresponding to a serological model object
containing information about the original serological survey data used to
fit the model, such as the year when the survey took place, the type of test
taken and the corresponding antibody, as well as information about the
convergence of the model, like the expected log pointwise predictive density
\code{elpd} and its corresponding standard deviation.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_object <- run_seromodel(
  serodata = serodata,
  foi_model = "constant"
)
extract_seromodel_summary(seromodel_object,
  serodata = serodata
)
}
