% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{extract_summary_model}
\alias{extract_summary_model}
\title{Extract Summary Model}
\usage{
extract_summary_model(model_object)
}
\arguments{
\item{model_object}{what the run model function returns}

\item{model_data}{refers to data of the model}
}
\value{
summary of the models
}
\description{
Function that summarizes the models
}
