% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_model}
\alias{fit_model}
\title{Fit Model}
\usage{
fit_model(
  model_data,
  model_name,
  n_iters = 1000,
  n_thin = 2,
  delta = 0.9,
  m_treed = 10,
  decades = 0
)
}
\arguments{
\item{model_data}{A data frame containing the data from a seroprevalence survey. For further details refer to \link{run_model}.}

\item{model_name}{Name of the selected model. Current version provides three options:
\describe{
\item{\code{constant_foi_bi}}{Runs a constant model}
\item{\code{continuous_foi_normal_bi}}{Runs a normal model}
\item{\code{continuous_foi_normal_log}}{Runs a normal logarithmic model}
}}

\item{n_iters}{Number of interations for eah chain including the warmup. \code{iter} in \link[rstan]{sampling}.}

\item{n_thin}{Positive integer specifying the period for saving samples. \code{thin} in \link[rstan]{sampling}.}

\item{delta}{Real number between 0 and 1 that represents the target average acceptance probability.
Increasing the value of \code{delta} will result in a smaller step size and fewer divergences.
For further details refer to the \code{control} parameter in \link[rstan]{sampling} or \href{https://mc-stan.org/rstanarm/reference/adapt_delta.html}{here}.}

\item{m_treed}{Maximum tree depth for the binary tree used in the NUTS stan sampler. For further details refer to the \code{control} parameter in \link[rstan]{sampling}.}

\item{decades}{Number of decades covered by the survey data.}
}
\value{
model_object
}
\description{
Function that fits the selected model to the data
}
\examples{
model_data <- prepare_data(mydata)
fit_model (model_data,
           model_name = "constant_foi_bi",
           n_iters = n_iters,
           n_thin = n_thin,
           delta = delta,
           m_treed = m_treed,
           decades = decades)
}
