% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_model}
\alias{fit_model}
\title{Fit Model}
\usage{
fit_model(
  model,
  model_data,
  model_name,
  n_iters = n_iters,
  n_thin = n_thin,
  delta = delta,
  m_treed = m_treed,
  decades = decades
)
}
\arguments{
\item{model}{refers to model selected}

\item{model_data}{refers to the model data that has been selected}

\item{model_name}{name of the model selected}

\item{n_iters}{number of iterations. Each model has a default number.}

\item{n_thin}{Each model has a default number.}

\item{delta}{This value comes by default but it can be changed}

\item{m_treed}{This value comes by default but it can be changed}

\item{decades}{The decades covered by the survey data}
}
\value{
model_object
}
\description{
Function that fits the model to the data
}
