% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{fit_model}
\alias{fit_model}
\title{Fit Model}
\usage{
fit_model(
  model,
  model_data,
  model_name,
  n_iters = n_iters,
  n_thin = n_thin,
  delta = delta,
  m_treed = m_treed,
  decades = decades
)
}
\arguments{
\item{model_data}{model_data}
}
\value{
model_object
}
\description{
Función que ajusta el modelo a los datos
Function that fits the model to the data
}
