% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{fit_model_log}
\alias{fit_model_log}
\title{Fit Model Log}
\usage{
fit_model_log(
  model,
  model_data,
  model_name,
  n_iters = 3000,
  n_thin = 2,
  delta = 0.9,
  m_treed = 10,
  decades = 0
)
}
\arguments{
\item{model_data}{model_data}
}
\value{
model_object
}
\description{
Función que ajusta el modelo logarítmico a los datos
Function that fits the logarithmic model to the data
}
