% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{generate_sim_data}
\alias{generate_sim_data}
\title{Function that generates a simulated serosurvey according to the specified FoI}
\usage{
generate_sim_data(
  foi,
  sample_size_by_age,
  tsur,
  birth_year_min,
  survey_label,
  test = "fake",
  antibody = "IgG",
  seed = 1234
)
}
\arguments{
\item{foi}{Numeric atomic vector corresponding to the desired Force-of-Infection ordered from past to present}

\item{sample_size_by_age}{Size of each age group specified by either an atomic
vector of the same size as \code{foi} or an integer.
This corresponds to the number of trials \code{size} in \link[stats]{rbinom}.}
}
\value{
Dataframe object containing the simulated data generated from \code{foi}
}
\description{
Function that generates a simulated serosurvey according to the specified FoI
}
\examples{
\dontrun{
sample_size_by_age = 5
foi <- rep(0.02, 50)
sim_data <- generate_sim_data(foi = foi,
                              sample_size_by_age = sample_size_by_age,
                              tsur = 2050,
                              birth_year_min = 2000,
                              survey_label = 'sim_constant_foi')
}
}
