% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{get_age_group}
\alias{get_age_group}
\title{Method for constructing age-group variable from age column}
\usage{
get_age_group(data, col_age, max_val, min_val, step)
}
\arguments{
\item{data}{dataframe with at least a column containing the age
information}

\item{col_age}{name of the column containing the age
information}

\item{max_val}{maximum value of age interval to split}

\item{min_val}{minimum value of age interval to split}

\item{step}{step used to split the age interval}
}
\value{
age_group factor variable grouping \code{data} by the age intervals 
specified by \code{min_val}, \code{max_val} and \code{col_age}
}
\description{
This function was taken from \link[vaccineff]{get_age_group}.
This method splits an age interval from min_val to max_val into
(max_val-min_val)/step intervals.
By default min_val is set 0, however it can be assigned by
convenience.
If the method finds ages greater or equal than max_val
it assigns the string ">{max_val}".
To avoid errors it is necessary to set step < max_val.
It is also suggested to choose the step such
that max_val%%(step+1) = 0.
}
