% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{get_age_group}
\alias{get_age_group}
\title{Method for constructing age-group variable from age column}
\usage{
get_age_group(age, step)
}
\arguments{
\item{age}{vector containing age information}

\item{step}{step used to split the age interval}
}
\value{
age_group factor variable grouping \code{age} by the age intervals
specified by \code{min(age)}, \code{max(age)}.
}
\description{
This function was taken from \link[vaccineff:get_age_group]{get_age_group}.
This method splits an age interval from age_min to age_max into
(age_max-age_min)/step intervals.
By default age_min is set 0, however it can be assigned by
convenience.
If the method finds ages greater or equal than age_max
it assigns the string ">{age_max}".
To avoid errors it is necessary to set step < age_max.
It is also suggested to choose the step such
that age_max\%\%(step+1) = 0.
}
