% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_cohort_ages}
\alias{get_cohort_ages}
\title{Function that generates a data.frame containing the age of each cohort corresponding to each birth year exluding the year of the survey.}
\usage{
get_cohort_ages(serodata)
}
\arguments{
\item{serodata}{A data frame containing the data from a seroprevalence survey.
This data frame must contain the following columns:
\tabular{ll}{
\code{survey} \tab survey Label of the current survey \cr \tab \cr
\code{total} \tab Number of samples for each age group\cr \tab \cr
\code{counts} \tab Number of positive samples for each age group\cr \tab \cr
\code{age_min} \tab age_min \cr \tab \cr
\code{age_max} \tab age_max \cr \tab \cr
\code{tsur} \tab Year in which the survey took place \cr \tab \cr
\code{country} \tab The country where the survey took place \cr \tab \cr
\code{test} \tab The type of test taken \cr \tab \cr
\code{antibody} \tab antibody \cr \tab \cr
\code{age_mean_f} \tab Floor value of the average between age_min and age_max \cr \tab \cr
\code{sample_size} \tab The size of the sample \cr \tab \cr
\code{birth_year} \tab The year in which the individuals of each age group were bornt \cr \tab \cr
\code{prev_obs} \tab Observed prevalence \cr \tab \cr
\code{prev_obs_lower} \tab Lower limit of the confidence interval for the observed prevalence \cr \tab \cr
\code{prev_obs_upper} \tab Upper limit of the confidence interval for the observed prevalence \cr \tab \cr
}
The last six colums can be added to \code{serodata} by means of the function \code{\link{prepare_serodata}}.}
}
\value{
\code{cohort_ages}. A data.frame containing the age of each cohort corresponding to each birth year
}
\description{
This function generates a data.frame containing the age of each cohort corresponding to each \code{birth_year} excluding the year of the survey,
for which the cohort age is still 0.
specified serological survey data \code{serodata} excluding the year of the survey.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(serodata = chagas2012, alpha = 0.05)
cohort_ages <- get_cohort_ages(serodata = serodata)
}
