% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_cohort_ages}
\alias{get_cohort_ages}
\title{Function that generates a data.frame containing the age of each cohort
corresponding to each birth year excluding the year of the survey.}
\usage{
get_cohort_ages(serodata)
}
\arguments{
\item{serodata}{A data frame containing the data from a seroprevalence
survey. This data frame must contain at least the following columns:
\describe{
\item{\code{total}}{Number of samples for each age group}
\item{\code{counts}}{Number of positive samples for each age group}
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{age_mean_f}}{Floor value of the average between age_min and age_max}
\item{\code{sample_size}}{The size of the sample}
\item{\code{birth_year}}{The year in which the individuals of each age group
were born}
}
The last six columns can be added to \code{serodata} by means of the function
\code{\link[=prepare_serodata]{prepare_serodata()}}.}
}
\value{
\code{cohort_ages}. A data.frame containing the age of each cohort
corresponding to each birth year
}
\description{
This function generates a data.frame containing the age of each cohort
corresponding to each \code{birth_year} excluding the year of the survey, for
which the cohort age is still 0. specified serological survey data \code{serodata}
excluding the year of the survey.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(serodata = chagas2012, alpha = 0.05)
cohort_ages <- get_cohort_ages(serodata = serodata)
}
