% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{get_comparison_table}
\alias{get_comparison_table}
\title{Get Model Table Comparison
Provides a table with statistics for comparison between models and selection}
\usage{
get_comparison_table(model_objects_list)
}
\arguments{
\item{model_objects_list}{model_objects to compare}
}
\value{
comparison table
}
\description{
Get Model Table Comparison
Provides a table with statistics for comparison between models and selection
}
