% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_exposure_ages}
\alias{get_exposure_ages}
\title{Function that generates an atomic vector containing the corresponding exposition years of a serological survey}
\usage{
get_exposure_ages(serodata)
}
\arguments{
\item{serodata}{A data frame containing the data from a seroprevalence survey. This data frame must contain the year of birth for each individual (birth_year) and the time of the survey (tsur). birth_year can be constructed by means of the \link{prepare_serodata} function.}
}
\value{
\code{exposure_ages}. An atomic vector with the numeration of the exposition years in serodata
}
\description{
This function generates an atomic vector containing the exposition years corresponding to the specified serological survey data \code{serodata}.
The exposition years to the disease for each individual corresponds to the time from birth to the moment of the survey.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(serodata = chagas2012, alpha = 0.05)
exposure_ages <- get_exposure_ages(serodata)
}
