% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_force.R
\name{get_exposure_matrix}
\alias{get_exposure_matrix}
\title{Dir Results}
\usage{
get_exposure_matrix(dat, yexpo)
}
\value{
dir
}
\description{
Función guarda los resultados en una dirección
Function saves the results in a place
}
