% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_exposure_matrix}
\alias{get_exposure_matrix}
\title{Get Exposure Matrix}
\usage{
get_exposure_matrix(model_data, yexpo)
}
\arguments{
\item{model_data}{refers to the model data that has been selected}

\item{yexpo}{what the make yexpo function returns}
}
\value{
exposure_output
}
\description{
Function that gets the exposure matrix
}
