% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_exposure_matrix}
\alias{get_exposure_matrix}
\title{Function that generates the exposure matrix corresponding to a serological survey}
\usage{
get_exposure_matrix(serodata)
}
\arguments{
\item{serodata}{A data frame containing the data from a seroprevalence survey.
This data frame must contain the year of birth for each individual (birth_year) and the time of the survey (tsur).
birth_year can be constructed by means of the \link{prepare_serodata} function.}
}
\value{
\code{exposure_output}. An atomic matrix containing the expositions for each entry of \code{serodata} by year.
}
\description{
Function that generates the exposure matrix corresponding to a serological survey
}
\examples{
\dontrun{
data("serodata")
serodata <- prepare_serodata(serodata = serodata)
exposure_matrix <- get_exposure_matrix(serodata = serodata)
}
}
