% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_foi_central_estimates}
\alias{get_foi_central_estimates}
\title{Function that generates the central estimates for the fitted forced FoI}
\usage{
get_foi_central_estimates(seromodel_object)
}
\arguments{
\item{seromodel_object}{Object containing the results of fitting a model by means of \link{run_seromodel}.
generated by means of \link{get_exposure_ages}.}
}
\value{
\code{foi_central_estimates}. Central estimates for the fitted forced FoI
}
\description{
Function that generates the central estimates for the fitted forced FoI
}
\examples{
\dontrun{
data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_object <- fit_seromodel(serodata = serodata,
                                  foi_model = "constant")
foi_central_estimates <- get_foi_central_estimates(seromodel_object)
}

}
