% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_foi_central_estimates}
\alias{get_foi_central_estimates}
\title{Function that generates the central estimates for the fitted forced FoI}
\usage{
get_foi_central_estimates(seromodel_object, cohort_ages)
}
\arguments{
\item{seromodel_object}{Stanfit object containing the results of fitting a model by means of \link{run_seromodel}.
generated by means of \link{get_exposure_ages}.}

\item{cohort_ages}{A data.frame containing the age of each cohort corresponding to each birth year.}
}
\value{
\code{foi_central_estimates}. Central estimates for the fitted forced FoI
}
\description{
Function that generates the central estimates for the fitted forced FoI
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_object <- fit_seromodel(serodata = serodata,
                                  foi_model = "constant")
cohort_ages <- get_cohort_ages(serodata = serodata)
foi_central_estimates <- get_foi_central_estimates(seromodel_object = seromodel_object, 
                                                   cohort_ages = cohort_ages)
}
