% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{get_model_comparison_plot}
\alias{get_model_comparison_plot}
\title{Get Model Comparison Plot}
\usage{
get_model_comparison_plot(result_comp)
}
\arguments{
\item{result_comp}{what the compare and save best model function returns}

\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
a model comparison graph
}
\description{
Function that obtains a model comparison plot
}
