% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_posterior_summary}
\alias{get_posterior_summary}
\title{Get Posterior Summary}
\usage{
get_posterior_summary(model_objects_chain)
}
\arguments{
\item{model_objects}{model_objects_chain}
}
\value{
model_object
}
\description{
Function that gets a posterior summary
}
\examples{
get_posterior_summary (model_objects_chain)
}
