% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_force.R
\name{get_posterior_summary}
\alias{get_posterior_summary}
\title{Get Posterior Summary}
\usage{
get_posterior_summary(results_chain)
}
\arguments{
\item{results}{results_chain}
}
\value{
results
}
\description{
Función que consigue un resumen posterior
Function that gets a post summary
}
