% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{get_prev_expanded}
\alias{get_prev_expanded}
\title{Get Prevalence Expanded}
\usage{
get_prev_expanded(foi, model_data)
}
\arguments{
\item{model_data}{model_data}
}
\value{
prev_final
}
\description{
Función que obtiene la prevalencia expandida
Function that obtains the expanded prevalence
}
