% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_prev_expanded}
\alias{get_prev_expanded}
\title{Get Prevalence Expanded}
\usage{
get_prev_expanded(foi, serodata)
}
\arguments{
\item{foi}{Object containing the information of the force of infection. It is obtained from \code{rstan::extract(seromodel_object$fit, "foi", inc_warmup = FALSE)[[1]]}.}

\item{serodata}{A data frame containing the data from a seroprevalence survey. For further details refer to \link{run_seromodel}.}
}
\value{
\code{prev_final}. The expanded prevalence data. This is used for plotting purposes in the \code{visualization} module.
}
\description{
Function that generates the expanded prevalence
}
\examples{
\dontrun{
serodata <- prepare_serodata(serodata)
seromodel_object <- run_seromodel(serodata = serodata,
                          seromodel_name = "constant_foi_bi")
foi <- rstan::extract(seromodel_object$fit, "foi", inc_warmup = FALSE)[[1]]
get_prev_expanded <- function(foi, serodata)
}
}
