% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_prev_expanded}
\alias{get_prev_expanded}
\title{Function that generates an object containing the confidence interval based on
a Force-of-Infection fitting}
\usage{
get_prev_expanded(
  foi,
  serodata,
  predicted_prev_lower_quantile = 0.1,
  predicted_prev_upper_quantile = 0.9,
  bin_data = FALSE
)
}
\arguments{
\item{foi}{Object containing the information of the force of infection. It is
obtained from \code{rstan::extract(seromodel_object$seromodel, "foi", inc_warmup = FALSE)[[1]]}.}

\item{serodata}{A data frame containing the data from a seroprevalence
survey. This data frame must contain the following columns:
\describe{
\item{\code{survey}}{survey Label of the current survey}
\item{\code{total}}{Number of samples for each age group}
\item{\code{counts}}{Number of positive samples for each age group}
\item{\code{age_min}}{age_min}
\item{\code{age_max}}{age_max}
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{country}}{The country where the survey took place}
\item{\code{test}}{The type of test taken}
\item{\code{antibody}}{antibody}
\item{\code{age_mean_f}}{Floor value of the average between age_min and age_max}
\item{\code{sample_size}}{The size of the sample}
\item{\code{birth_year}}{The year in which the individuals of each age group
were born}
\item{\code{prev_obs}}{Observed prevalence}
\item{\code{prev_obs_lower}}{Lower limit of the confidence interval for the
observed prevalence}
\item{\code{prev_obs_upper}}{Upper limit of the confidence interval for the
observed prevalence}
}
The last six columns can be added to \code{serodata} by means of the function
\code{\link[=prepare_serodata]{prepare_serodata()}}.}

\item{predicted_prev_lower_quantile}{Float specifying the lower quantile
to be used in \code{predicted_prev} calculation}

\item{predicted_prev_upper_quantile}{Float specifying the upper quantile
to be used in \code{predicted_prev} calculation}

\item{bin_data}{If \code{TRUE}, \code{serodata} is binned by means of
\code{prepare_bin_data}. Otherwise, age groups are kept as originally input.}
}
\value{
\code{prev_final}. The expanded prevalence data. This is used for plotting
purposes in the \code{visualization} module.
}
\description{
This function computes the corresponding binomial confidence intervals for
the obtained prevalence based on a fitting of the Force-of-Infection \code{foi}
for plotting an analysis purposes.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_object <- run_seromodel(
  serodata = serodata,
  foi_model = "constant"
)
foi <- rstan::extract(seromodel_object, "foi")[[1]]
get_prev_expanded(foi, serodata)
}
