% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_prev_expanded}
\alias{get_prev_expanded}
\title{Get Prevalence Expanded}
\usage{
get_prev_expanded(foi, model_data)
}
\arguments{
\item{foi}{force of infection}

\item{model_data}{refers to the model data that has been selected}
}
\value{
prev_final
}
\description{
Function that generates the expanded prevalence
}
\examples{
get_prev_expanded <- function(foi, model_data)
}
