% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_prev_expanded}
\alias{get_prev_expanded}
\title{Generate object containing the confidence interval based on
a force-of-infection fitting}
\usage{
get_prev_expanded(foi, serodata, alpha = 0.05, bin_data = FALSE, bin_step = 5)
}
\arguments{
\item{foi}{Object containing the information of the force-of-infection. It is
obtained from \code{rstan::extract(seromodel_object$seromodel, "foi", inc_warmup = FALSE)[[1]]}.}

\item{serodata}{A data frame containing the data from a seroprevalence
survey. This data frame must contain at least the following columns:
\describe{
\item{\code{total}}{Number of samples for each age group}
\item{\code{counts}}{Number of positive samples for each age group}
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{age_mean_f}}{Floor value of the average between age_min and age_max}
\item{\code{sample_size}}{The size of the sample}
\item{\code{birth_year}}{The year in which the individuals of each age group
were born}
}
The last six columns can be added to \code{serodata} by means of the function
\code{\link[=prepare_serodata]{prepare_serodata()}}.}

\item{alpha}{Probability threshold for statistical significance used for both
the binomial confidence interval, and the lower and upper quantiles of the
estimated prevalence.}

\item{bin_data}{Boolean. Use \code{TRUE} when age binning is preferred for plotting.
If \code{TRUE}, \code{serodata} is binned by means of \code{prepare_bin_data};
Otherwise, age groups are kept as originally input.}

\item{bin_step}{Integer specifying the age groups bin size to be used when
\code{bin_data} is set to \code{TRUE}.}
}
\value{
\code{prev_final}. The expanded prevalence data. This is used for plotting
purposes in the \code{visualization} module.
}
\description{
This function computes the corresponding binomial confidence intervals for
the obtained prevalence based on a fitting of the force-of-infection \code{foi}
for plotting an analysis purposes.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_object <- run_seromodel(
  serodata = serodata,
  foi_model = "constant"
)
foi <- rstan::extract(seromodel_object, "foi")[[1]]
get_prev_expanded(foi, serodata)
}
