% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_prev_expanded}
\alias{get_prev_expanded}
\title{Function that generates an object containing the confidence interval based on a
Force-of-Infection fitting}
\usage{
get_prev_expanded(foi, serodata, bin_data = FALSE)
}
\arguments{
\item{foi}{Object containing the information of the force of infection. It is obtained from \code{rstan::extract(seromodel_object$seromodel, "foi", inc_warmup = FALSE)[[1]]}.}

\item{serodata}{A data frame containing the data from a seroprevalence survey. For further details refer to \link{run_seromodel}.}

\item{bin_data}{TBD}
}
\value{
\code{prev_final}. The expanded prevalence data. This is used for plotting purposes in the \code{visualization} module.
}
\description{
This function computes the corresponding binomial confidence intervals for the obtained prevalence based on a fitting
of the Force-of-Infection \code{foi} for plotting an analysis purposes.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_object <- run_seromodel(serodata = serodata,
                          foi_model = "constant")
foi <- rstan::extract(seromodel_object$seromodel_fit, "foi")[[1]]
get_prev_expanded(foi, serodata)
}
