% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{get_residuals}
\alias{get_residuals}
\title{Get Residuals}
\usage{
get_residuals(fit, model_data)
}
\arguments{
\item{model_data}{model_data}
}
\value{
merged_prev
}
\description{
Función que obtiene los residuos
Function that gets the residuals
}
