% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_residuals}
\alias{get_residuals}
\title{Get Residuals}
\usage{
get_residuals(fit, model_data)
}
\arguments{
\item{fit}{refers to fit of the model}

\item{model_data}{refers to the model data that has been selected}
}
\value{
merged_prev
}
\description{
Function that gets the residuals
}
