% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{get_sim_counts}
\alias{get_sim_counts}
\title{Function that randomly generates a sample of counts for a simulated dataset}
\usage{
get_sim_counts(sim_data, foi, size_age_class, seed = 1234)
}
\arguments{
\item{sim_data}{A dataframe object containing the following columns:
\tabular{ll}{
\code{birth_year} \tab List of years in which the subjects were borned \cr \tab \cr
\code{tsur} \tab Year of the survey\cr \tab \cr
\code{country} \tab Default to 'none'.\cr \tab \cr
\code{survey} \tab Survey label \cr \tab \cr
\code{age_mean_f} \tab Age \cr \tab \cr
}}
}
\value{
A simulated list of counts following a binomial distribution in accordance with a given force of infection and age class sizes.
}
\description{
Function that randomly generates a sample of counts for a simulated dataset
}
\examples{
\dontrun{

}
}
