% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{get_sim_n_seropositive}
\alias{get_sim_n_seropositive}
\title{Generate sample of counts of seropositive individuals by
sampling from a binomial distribution}
\usage{
get_sim_n_seropositive(
  sim_data,
  foi,
  sample_size_by_age,
  mu = 0,
  model_type = "time-varying",
  seed = 1234
)
}
\arguments{
\item{sim_data}{A dataframe object containing the following columns:
\describe{
\item{\code{age_mean_f}}{Age group markers}
\item{\code{tsur}}{Year of the survey}
}}

\item{foi}{Numeric atomic vector corresponding to the desired
time-varying or age-varying Force-of-Infection to simulate from}

\item{sample_size_by_age}{Integer indicating the sample size by age group.
This corresponds to the number of trials \code{size} in \link[stats:Binomial]{rbinom}.}

\item{mu}{Seroreversion rate}

\item{model_type}{String specifying the type of model to be used.
Current valid options are 'time-varying' and 'age-varying'}

\item{seed}{Seed for random number generation.}
}
\value{
A dataframe containing the following columns:
\describe{
\item{\code{age}}{Age by the time of the survey}
\item{\code{n_seropositive}}{Number of positive cases sampled according to the
provided FoI}
}
simulated list of counts following a binomial distribution in accordance
with a given force of infection and age group sizes.
}
\description{
Generate sample of counts of seropositive individuals by
sampling from a binomial distribution
}
\examples{
n_years <- 50
sim_data <- data.frame(
  age_mean_f = seq(1,n_years),
  tsur = 2050
)
foi <- rep(0.02, n_years)
sample_size_by_age <- as.integer(runif(n = n_years, 5, 10))
sim_n_seropositive <- get_sim_n_seropositive(
  sim_data = sim_data,
  foi = foi,
  sample_size_by_age = sample_size_by_age
)
}
