% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{get_sim_probability}
\alias{get_sim_probability}
\title{Generate probabilities of being previously exposed to a
pathogen given a historical force-of-infection.}
\usage{
get_sim_probability(
  sim_data,
  foi,
  mu = 0,
  is_time_varying = TRUE,
  is_age_varying = FALSE
)
}
\arguments{
\item{sim_data}{A dataframe object containing the following columns:
\describe{
\item{\code{age_mean_f}}{Age group markers}
\item{\code{tsur}}{Year of the survey}
}}

\item{foi}{Numeric atomic vector corresponding to the desired
force-of-infection ordered from past to present}
}
\value{
A dataframe containing the following columns:
\describe{
\item{\code{age}}{Exposure ages}
\item{\code{probability}}{Probability to obtain a seropositive case for each age
according to the provided FoI}
}
}
\description{
Generate probabilities of being previously exposed to a
pathogen given a historical force-of-infection.
}
\examples{
n_years <- 50
sim_data <- data.frame(
  age_mean_f = seq(1,n_years),
  tsur = 2050
)
foi <- rep(0.02, n_years)
sim_probability <- get_sim_probability(sim_data = sim_data, foi=foi)
}
