% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{get_sim_probability}
\alias{get_sim_probability}
\title{Function that generates the probabilities of being previously exposed to a pathogen.}
\usage{
get_sim_probability(sim_data, foi)
}
\arguments{
\item{sim_data}{A dataframe object containing the following columns:
\tabular{ll}{
\code{tsur} \tab Year of the survey\cr \tab \cr
\code{age_mean_f} \tab Age \cr \tab \cr
\code{birth_year} \tab Years in which the subjects were borned
according to the age group marker \code{age_mean_f}\cr \tab \cr
}}

\item{foi}{Numeric atomic vector corresponding to the desired Force-of-Infection ordered from past to present}

\item{seed}{The seed for random number generation.}
}
\value{
A dataframe containing the following columns:
\tabular{ll}{
\code{age} \tab Exposure ages \cr \tab \cr
\code{probability} \tab Probability to obtain a seropositive case for each age according to the provided FoI\cr \tab \cr
}
}
\description{
Function that generates the probabilities of being previously exposed to a pathogen.
}
