% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{get_table_rhats}
\alias{get_table_rhats}
\title{Method for extracting a dataframe containing the R-hat estimates for a given serological model}
\usage{
get_table_rhats(seromodel_object, serodata)
}
\arguments{
\item{seromodel_object}{seromodel_object}
}
\value{
rhats table
}
\description{
This method relies in the function \link[bayesplot]{rhat} to extract the R-hat estimates of the serological model object 
\code{seromodel_object} and returns a table a dataframe with the estimates for each year of birth.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(serodata = chagas2012)
model_constant <- run_seromodel(serodata = serodata,
                                foi_model = "constant",
                                n_iters = 1500)
get_table_rhats(seromodel_object = model_constant,
                serodata = serodata)
}
