% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{get_table_rhats}
\alias{get_table_rhats}
\title{Get Table Rhats}
\usage{
get_table_rhats(res)
}
\value{
rhats table
}
\description{
Función que hace la tabla de los rhats
Function that makes the rhats table
}
