% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{get_table_rhats}
\alias{get_table_rhats}
\title{Get Table Rhats}
\usage{
get_table_rhats(model_object)
}
\arguments{
\item{model_object}{model_object}
}
\value{
rhats table
}
\description{
Function that makes the rhats table
}
\examples{
model_data <- prepare_data(model_data = mydata, alpha = 0.05)
model_object <- run_model(
model_data = model_data, model_name = "constant_foi_bi")
get_table_rhats (model_object)
}
