% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{get_table_rhats}
\alias{get_table_rhats}
\title{Method for extracting a dataframe containing the R-hat estimates for a given serological model}
\usage{
get_table_rhats(seromodel_object)
}
\arguments{
\item{seromodel_object}{seromodel_object}
}
\value{
rhats table
}
\description{
This method relies in the function \link[bayesplot]{rhat} to extract the R-hat estimates of the serological model object 
\code{seromodel_object} and returns a table a dataframe with the estimates for each year of birth.
}
\examples{
\dontrun{
data("serodata")
data_test <- prepare_serodata(serodata = serodata)
model_constant <- run_seromodel(serodata = data_test, 
                                seromodel_name = "constant_foi_bi", 
                                n_iters = 1500)
get_table_rhats(model_object = model_constant)
}
}
