% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{get_table_rhats}
\alias{get_table_rhats}
\title{Method for extracting a dataframe containing the R-hat estimates for a given serological model}
\usage{
get_table_rhats(seromodel_object, cohort_ages)
}
\arguments{
\item{seromodel_object}{Stanfit object containing the results of fitting a model by means of \link{run_seromodel}.}

\item{cohort_ages}{A data.frame containing the age of each cohort corresponding to each birth year.}
}
\value{
rhats table
}
\description{
This method relies in the function \link[bayesplot]{rhat} to extract the R-hat estimates of the serological model object 
\code{seromodel_object} and returns a table a dataframe with the estimates for each year of birth.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(serodata = chagas2012)
model_constant <- run_seromodel(serodata = serodata,
                                foi_model = "constant",
                                n_iters = 1500)
cohort_ages <- get_cohort_ages(serodata)
get_table_rhats(seromodel_object = model_constant,
                cohort_ages = cohort_ages)
}
