% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{get_table_rhats}
\alias{get_table_rhats}
\title{Get Table Rhats}
\usage{
get_table_rhats(model_object)
}
\arguments{
\item{model_object}{model_object}
}
\value{
rhats table
}
\description{
Function that makes the rhats table
}
\examples{
\dontrun{
seroprev_data <- prepare_seroprev_data(seroprev_data = mydata, alpha = 0.05)
model_object <- run_seroprev_model(
seroprev_data = seroprev_data, seroprev_model_name = "constant_foi_bi")
get_table_rhats (model_object)
}
}
