% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_force.R
\name{make_thin_chain}
\alias{make_thin_chain}
\title{Make Thin Chain}
\usage{
make_thin_chain(results_chain, thin = 10)
}
\value{
results_chain
}
\description{
Función que hace el adelgazamiento del número de iteraciones
Function that does the thinning of the number of iterations
}
