% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{make_yexpo}
\alias{make_yexpo}
\title{Make Yexpo}
\usage{
make_yexpo(model_data)
}
\arguments{
\item{model_data}{model_data}
}
\value{
Yexpo
}
\description{
Función que hace Yexpo
Function thats make Yexpo
}
