% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{obtain_prevalence_extended}
\alias{obtain_prevalence_extended}
\title{Obtain Prevalence Extended}
\usage{
obtain_prevalence_extended(model_data, exposure, ly, nbreaks, lambdaYexpo)
}
\arguments{
\item{model_data}{refers to the model data that has been selected}
}
\value{
new_PPP
}
\description{
Function that obtains the extended prevalence
}
