% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_foi}
\alias{plot_foi}
\title{Function that generates a Force-of-Infection plot corresponding to the specified fitted serological model}
\usage{
plot_foi(seromodel_object, lambda_sim = NA, max_lambda = NA, size_text = 25)
}
\arguments{
\item{seromodel_object}{Object containing the results of fitting a model by means of \link{run_seromodel}.}

\item{size_text}{Text size use in the theme of the graph returned by the function.}
}
\value{
A ggplot2 object containing the Force-of-infection-vs-time including the corresponding confidence interval.
}
\description{
This function generates a Force-of-Infection plot from the results obtained by fitting a serological model.
This includes the corresponding binomial confidence interval. 
The x axis corresponds to the decades covered by the survey the y axis to the Force-of-Infection.
}
\examples{
\dontrun{
   data_test <- prepare_serodata(serodata)
   seromodel_object <- run_seromodel(
   serodata = data_test,
   foi_model = "constant",
   n_iters = 1000
)
plot_foi(seromodel_object, size_text = 15)
}
}
