% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_foi}
\alias{plot_foi}
\title{Generate Force-of-Infection Plot}
\usage{
plot_foi(model_object, lambda_sim = NA, max_lambda = NA, size_text = 25)
}
\arguments{
\item{model_object}{Object containing the results of fitting a model by means of \link{run_seroprev_model}.}

\item{size_text}{Text size use in the theme of the graph returned by the function.}
}
\value{
A ggplot2 object containing the Force-of-infection-vs-time including the corresponding confidence interval.
}
\description{
Function that generates the Force-of-Infection plot. The x axis corresponds to the decades covered by the survey the y axis to the Force-of-Infection.
}
\examples{
\dontrun{
   data_test <- preprare_seroprev_data(mydata)
   model_object <- run_seroprev_model(
   seroprev_data = data_test,
   seroprev_model_name = "constant_foi_bi",
   n_iters = 1000
)
plot_foi(model_object, size_text = 15)
}
}
