% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_foi}
\alias{plot_foi}
\title{Generate Force-of-Infection Plot}
\usage{
plot_foi(model_object, lambda_sim = NA, max_lambda = NA, size_text = 25)
}
\arguments{
\item{model_object}{Object that the run_model function returns with the results of the fit}

\item{model}{Refers to model selected}

\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
Force of infection plot
}
\description{
Function that generates the force of infection plot
}
\examples{
 data_test <- prepare_data(mydata)
model_object <- run_model(
  model_data = data_test,
  model_name = "constant_foi_bi",
  n_iters = 1000
)
plot_foi(model_object, size_text = 15)
}
