% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_foi}
\alias{plot_foi}
\title{Generate Force-of-Infection Plot}
\usage{
plot_foi(model_object, lambda_sim = NA, max_lambda = NA, size_text = 25)
}
\arguments{
\item{model_object}{Object that the run_model function returns with the results of the fit}

\item{size_text}{Text size of the graph returned by the function}

\item{model}{It is the selected model that for the current version of serofoi are 3. For more information see the function \link{run_model}.}

\item{xlabel}{The years are located on the x-axis.}

\item{ylabel}{The force of infection is located on the y axis.}
}
\value{
The Force of infection plot with a 95% confidence interval.
}
\description{
Function that generates the force of infection plot
}
\examples{
\dontrun{
 data_test <- prepare_data(mydata)
model_object <- run_model(
  model_data = data_test,
  model_name = "constant_foi_bi",
  n_iters = 1000
)
plot_foi(model_object, size_text = 15)
}
}
