% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_foi}
\alias{plot_foi}
\title{Generate Force-of-Infection Plot}
\usage{
plot_foi(model_object, lambda_sim = NA, max_lambda = NA, size_text = 25)
}
\arguments{
\item{model_object}{what the run model function returns}

\item{model}{refers to model selected}

\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
Force of infection plot
}
\description{
Function that generates the force of infection plot
}
