% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_info_table}
\alias{plot_info_table}
\title{Plot Info Table}
\usage{
plot_info_table(info, size_text)
}
\arguments{
\item{model_data}{model_data}
}
\value{
The previous expanded graphic
}
\description{
Función que genera la tabla de información
Function that generates the information table
}
