% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_info_table}
\alias{plot_info_table}
\title{Generate Info Table Plot}
\usage{
plot_info_table(info, size_text)
}
\arguments{
\item{data}{Data}
}
\value{
The previous expanded graphic
}
\description{
Función que genera la tabla de información
Function that generates the information table
}
