% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_info_table}
\alias{plot_info_table}
\title{Plot Info Table}
\usage{
plot_info_table(info, size_text)
}
\arguments{
\item{info}{the information that will be contained in the table}

\item{size_text}{text size}

\item{model_data}{refers to data of the each model}
}
\value{
The previous expanded graphic
}
\description{
Function that generates the information table
}
