% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_info_table}
\alias{plot_info_table}
\title{Plot Info Table}
\usage{
plot_info_table(info, size_text)
}
\arguments{
\item{info}{the information that will be contained in the table}

\item{size_text}{Text size of the graph returned by the function}
}
\value{
p, a variable that will be used in the \link{visualisation} module
}
\description{
Function that generates the information table
}
\examples{
\dontrun{
 data_test <- prepare_seroprev_data(mydata)
 model_object <- run_seroprev_model(
 seroprev_data = data_test,
 seroprev_model_name = "constant_foi_bi",
 n_iters = 1000
)
info = t(model_object$model_summary)
plot_info_table (info, size_text = 15)
}
}
