% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_info_table}
\alias{plot_info_table}
\title{Function that generates a plot for a given table}
\usage{
plot_info_table(info, size_text)
}
\arguments{
\item{info}{the information that will be contained in the table}

\item{size_text}{Text size of the graph returned by the function}
}
\value{
p the plot for the given table
}
\description{
Function that generates a plot for a given table
}
\examples{
serodata <- prepare_serodata(chagas2012)
seromodel_object <- run_seromodel(
  serodata = serodata,
  foi_model = "constant",
  n_iters = 1000
)
seromodel_summary <- extract_seromodel_summary(
  seromodel_object = seromodel_object,
  serodata = serodata
)
info <- t(seromodel_summary)
plot_info_table(info, size_text = 15)
}
