% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_model}
\alias{plot_model}
\title{Generate Combined Plots}
\usage{
plot_model(model_object, lambda_sim = NA, max_lambda = NA, size_text = 25)
}
\arguments{
\item{model}{model data}

\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
The combined plots
}
\description{
Función que genera la gráfica combinada
Function that generates the combined graph
}
