% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_model}
\alias{plot_model}
\title{Generate a vertical arrange of plots summarizing the results of the model implementation}
\usage{
plot_model(model_object, lambda_sim = NA, max_lambda = NA, size_text = 25)
}
\arguments{
\item{model_object}{Object that the run_model function returns with the results of the fit}

\item{model}{Refers to model selected}

\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
The combined plots
}
\description{
Function that generates the combined graph
}
\examples{
plot_model(model_0_object, size_text)
}
