% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_rhats}
\alias{plot_rhats}
\title{Generate Rhats-Convergence Plot}
\usage{
plot_rhats(model_object, size_text = 25)
}
\arguments{
\item{model_object}{Object that the run_model function returns with the results of the fit}

\item{model}{Refers to model selected}

\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
The rhats-convergence plot of the selected model
}
\description{
Function that generates the convergence graph of a model
}
\examples{
data_test <- prepare_data(mydata)
model_object <- run_model(
 model_data = data_test,
 model_name = "constant_foi_bi",
 n_iters = 1000
)
plot_rhats(model_object, size_text = 15)
}
