% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_rhats}
\alias{plot_rhats}
\title{Generate Rhats-Convergence Plot}
\usage{
plot_rhats(model_object, size_text = 25)
}
\arguments{
\item{model}{model data}

\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
The rhats-convergence plot of a model
}
\description{
Función que genera la gráfica de convergencias de un modelo
Function that generates the convergence graph of a model
}
