% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_rhats}
\alias{plot_rhats}
\title{Generate Rhats-Convergence Plot}
\usage{
plot_rhats(model_object, size_text = 25)
}
\arguments{
\item{model_object}{Object containing the results of fitting a model by means of \link{run_seroprev_model}.}

\item{size_text}{Text size use in the theme of the graph returned by the function.}
}
\value{
The rhats-convergence plot of the selected model.
}
\description{
Function that generates the convergence graph of a model. The x axis corresponds to the decades covered by the survey and the y axis to the value of the rhats. 
All rhats must be smaller than 1 to ensure convergence.
}
\examples{
\dontrun{
data_test <- preprare_seroprev_data(mydata)
model_object <- run_seroprev_model(
 seroprev_data = data_test,
 seroprev_model_name = "constant_foi_bi",
 n_iters = 1000
)
plot_rhats(model_object, size_text = 15)
}
}
