% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_seromodel}
\alias{plot_seromodel}
\title{Generate vertical arrangement of plots showing a summary of a
model, the estimated seroprevalence, the force-of-infection fit and the R-hat
estimates plots.}
\usage{
plot_seromodel(
  seromodel_object,
  serodata,
  alpha = 0.05,
  max_lambda = NA,
  size_text = 25,
  bin_data = TRUE,
  bin_step = 5,
  foi_sim = NULL
)
}
\arguments{
\item{seromodel_object}{Stanfit object containing the results of fitting a
model by means of \link{run_seromodel}.}

\item{serodata}{A data frame containing the data from a seroprevalence
survey. This data frame must contain at least the following columns:
\describe{
\item{\code{total}}{Number of samples for each age group}
\item{\code{counts}}{Number of positive samples for each age group}
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{age_mean_f}}{Floor value of the average between age_min and age_max}
\item{\code{sample_size}}{The size of the sample}
\item{\code{birth_year}}{The year in which the individuals of each age group
were born}
}
The last six columns can be added to \code{serodata} by means of the function
\code{\link[=prepare_serodata]{prepare_serodata()}}.}

\item{alpha}{Probability threshold for statistical significance used for both
the binomial confidence interval, and the lower and upper quantiles of the
estimated prevalence.}

\item{max_lambda}{TBD}

\item{size_text}{Text size use in the theme of the graph returned by the
function.}

\item{bin_data}{If \code{TRUE}, \code{serodata} is binned by means of
\code{prepare_bin_data}. Otherwise, age groups are kept as originally input.}

\item{bin_step}{Integer specifying the age groups bin size to be used when
\code{bin_data} is set to \code{TRUE}.}

\item{foi_sim}{TBD}
}
\value{
A ggplot object with a vertical arrange containing the
seropositivity, force of infection, and convergence plots.
}
\description{
Generate vertical arrangement of plots showing a summary of a
model, the estimated seroprevalence, the force-of-infection fit and the R-hat
estimates plots.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_object <- run_seromodel(
  serodata = serodata,
  foi_model = "constant",
  iter = 1000
)
plot_seromodel(seromodel_object,
  serodata = serodata,
  size_text = 15
)
}
