% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_seromodel}
\alias{plot_seromodel}
\title{Function that generates a vertical arrange of plots showing a summary of a model, the estimated seroprevalence,
the Force-of-Infection fit and the R-hat estimates plots.}
\usage{
plot_seromodel(
  seromodel_object,
  max_lambda = NA,
  size_text = 25,
  foi_sim = NULL
)
}
\arguments{
\item{seromodel_object}{Object containing the results of fitting a model by means of \link{run_seromodel}.}

\item{max_lambda}{TBD}

\item{size_text}{Text size use in the theme of the graph returned by the function.}

\item{foi_sim}{TBD}
}
\value{
A ggplot object with a vertical arrange containing the seropositivity, force of infection, and convergence plots.
}
\description{
Function that generates a vertical arrange of plots showing a summary of a model, the estimated seroprevalence,
the Force-of-Infection fit and the R-hat estimates plots.
}
\examples{
\dontrun{
data_test <- prepare_serodata(serodata)
seromodel_object <- run_seromodel(
 serodata = data_test,
 foi_model = "constant",
 n_iters = 1000
)
plot_seromodel(seromodel_object, size_text = 15)
}
}
