% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_seromodel}
\alias{plot_seromodel}
\title{Function that generates a vertical arrange of plots showing a summary of a given implemented model, as well as its corresponding
seroprevalence, Force-of-Infection and R-hat estimates plots.}
\usage{
plot_seromodel(
  seromodel_object,
  lambda_sim = NA,
  max_lambda = NA,
  size_text = 25
)
}
\arguments{
\item{seromodel_object}{Object containing the results of fitting a model by means of \link{run_seromodel}.}

\item{size_text}{Text size use in the theme of the graph returned by the function.}
}
\value{
A ggplot object with a vertical arrange containing the seropositivity, force of infection, and convergence plots.
}
\description{
Function that generates a vertical arrange of plots showing a summary of a given implemented model, as well as its corresponding
seroprevalence, Force-of-Infection and R-hat estimates plots.
}
\examples{
\dontrun{
data_test <- prepare_serodata(serodata)
seromodel_object <- run_seromodel(
 serodata = data_test,
 seromodel_name = "constant_foi_bi",
 n_iters = 1000
)
plot_seromodel(seromodel_object, size_text = 15)
}
}
