% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_seroprev}
\alias{plot_seroprev}
\title{Generate sero-positivity plot from raw data}
\usage{
plot_seroprev(model_data, size_text = 6)
}
\arguments{
\item{model_object}{Object that the run_model function returns with the results of the fit}

\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
The sero-positivity plot
}
\description{
Function that generates the sero positivity plot
}
\examples{
\dontrun{
data_test <- prepare_data(mydata)
model_object <- run_model(
 model_data = data_test,
 model_name = "constant_foi_bi",
 n_iters = 1000
)
plot_seroprev(model_object, size_text = 15)
}
}
