% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_seroprev}
\alias{plot_seroprev}
\title{Generate sero-positivity plot of raw data}
\usage{
plot_seroprev(data_test)
}
\arguments{
\item{data_test}{Object with raw data}

\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
The sero-positivity plot
}
\description{
Function that generates the sero positivity plot of raw data
}
