% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_seroprev}
\alias{plot_seroprev}
\title{Function that generates the sero-positivity plot from a raw serological survey dataset.}
\usage{
plot_seroprev(serodata, size_text = 6)
}
\arguments{
\item{serodata}{A data frame containing the data from a seroprevalence survey.
This data frame must contain the following columns:
\tabular{ll}{
\code{survey} \tab survey Label of the current survey \cr \tab \cr
\code{total} \tab Number of samples for each age group\cr \tab \cr
\code{counts} \tab Number of positive samples for each age group\cr \tab \cr
\code{age_min} \tab age_min \cr \tab \cr
\code{age_max} \tab age_max \cr \tab \cr
\code{year_init} \tab year_init \cr \tab \cr
\code{year_end} \tab year_end \cr \tab \cr
\code{tsur} \tab Year in which the survey took place \cr \tab \cr
\code{country} \tab The country where the survey took place \cr \tab \cr
\code{test} \tab The type of test taken \cr \tab \cr
\code{antibody} \tab antibody \cr \tab \cr
}}

\item{size_text}{Text size use in the theme of the graph returned by the function.}
}
\value{
A ggplot object containing the seropositivity-vs-age graph of the raw data of a given seroprevalence survey with its corresponging binomial confidence interval.
}
\description{
Function that generates the sero-positivity plot from a raw serological survey dataset.
}
\examples{
\dontrun{
 data_test <- prepare_serodata(serodata)
 seromodel_object <- run_seromodel(
 serodata = data_test,
 seromodel_name = "constant_foi_bi",
 n_iters = 1000
)
plot_seroprev(seromodel_object, size_text = 15)
}
}
