% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_seroprev}
\alias{plot_seroprev}
\title{Function that generates the seropositivity plot from a raw serological
survey dataset}
\usage{
plot_seroprev(serodata, size_text = 6, bin_data = TRUE, bin_step = 5)
}
\arguments{
\item{serodata}{A data frame containing the data from a serological survey.
This data frame must contain the following columns:
\describe{
\item{\code{survey}}{survey Label of the current survey}
\item{\code{total}}{Number of samples for each age group}
\item{\code{counts}}{Number of positive samples for each age group}
\item{\code{age_min}}{age_min}
\item{\code{age_max}}{age_max}
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{country}}{The country where the survey took place}
\item{\code{test}}{The type of test taken}
\item{\code{antibody}}{antibody}
}
Alternatively to \code{age_min} and \code{age_max}, the dataset could already include
the age group marker \code{age_mean_f}, representing the middle point between
\code{age_min} and \code{age_max}. If \code{afe_mean_f} is missing, it will be generated
by the function.}

\item{size_text}{Text size use in the theme of the graph returned by the
function.}

\item{bin_data}{Boolean. Use \code{TRUE} when age binning is preferred for plotting.
If \code{TRUE}, \code{serodata} is binned by means of \code{prepare_bin_data};
Otherwise, age groups are kept as originally input.}

\item{bin_step}{Integer specifying the age groups bin size to be used when
\code{bin_data} is set to \code{TRUE}.}
}
\value{
A ggplot object containing the seropositivity-vs-age graph of the raw
data of a given seroprevalence survey with its corresponding binomial
confidence interval.
}
\description{
Function that generates the seropositivity plot from a raw serological
survey dataset
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(chagas2012)
plot_seroprev(serodata, size_text = 15)
}
