% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_seroprev}
\alias{plot_seroprev}
\title{Generate sero-positivity plot}
\usage{
plot_seroprev(model_object, size_text = 25)
}
\arguments{
\item{model_object}{what the run model function returns}

\item{model}{refers to model selected}

\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
The sero-positivity plot
}
\description{
Function that generates the sero positivity plot
}
