% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_seroprev_fitted}
\alias{plot_seroprev_fitted}
\title{Generate sero-positivity plot with fitted model}
\usage{
plot_seroprev_fitted(model_object, size_text = 6)
}
\arguments{
\item{model_object}{Object that the run_model function returns with the results of the fit}

\item{model}{Refers to the model selected}

\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
The sero-positivity plot
}
\description{
Function that generates the sero positivity plot
}
