% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_seroprev_fitted}
\alias{plot_seroprev_fitted}
\title{Function that generates a seropositivity plot corresponding to the specified
fitted serological model}
\usage{
plot_seroprev_fitted(
  seromodel_object,
  serodata,
  predicted_prev_lower_quantile = 0.1,
  predicted_prev_upper_quantile = 0.9,
  size_text = 6
)
}
\arguments{
\item{seromodel_object}{Stanfit object containing the results of fitting a
model by means of \link{run_seromodel}.}

\item{serodata}{A data frame containing the data from a seroprevalence
survey. This data frame must contain at least the following columns:
\describe{
\item{\code{total}}{Number of samples for each age group}
\item{\code{counts}}{Number of positive samples for each age group}
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{age_mean_f}}{Floor value of the average between age_min and age_max}
\item{\code{sample_size}}{The size of the sample}
\item{\code{birth_year}}{The year in which the individuals of each age group
were born}
}
The last six columns can be added to \code{serodata} by means of the function
\code{\link[=prepare_serodata]{prepare_serodata()}}.}

\item{predicted_prev_lower_quantile}{Float specifying the lower quantile
to be used in \code{predicted_prev} calculation}

\item{predicted_prev_upper_quantile}{Float specifying the upper quantile
to be used in \code{predicted_prev} calculation}

\item{size_text}{Text size of the graph returned by the function.}
}
\value{
A ggplot object containing the seropositivity-vs-age graph including
the data, the fitted model and their corresponding confidence intervals.
}
\description{
This function generates a seropositivity plot of the specified serological
model object. This includes the original data grouped by age as well as the
obtained fitting from the model implementation. Age is located on the x axis
and seropositivity on the y axis with its corresponding confidence interval.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_object <- run_seromodel(
  serodata = serodata,
  foi_model = "constant",
  iter = 1000
)
plot_seroprev_fitted(seromodel_object,
  serodata = serodata,
  size_text = 15
)
}
