% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_seroprev_fitted}
\alias{plot_seroprev_fitted}
\title{Generate sero-positivity plot with fitted model}
\usage{
plot_seroprev_fitted(model_object, size_text = 6)
}
\arguments{
\item{model_object}{Object that the run_model function returns with the results of the fit}

\item{size_text}{Text size of the graph returned by the function}

\item{model}{It is the selected model that for the current version of serofoi are 3. For more information see the function \link{run_model}.}

\item{xlabel}{On the x axis is the age found in the serological survey}

\item{ylabel}{Seropositivity is located on the y axis.}
}
\value{
Seropositivity graph according to age with seropositivity at a 95% confidence interval.
}
\description{
Function that generates the sero positivity plot with fitted model
}
\examples{
\dontrun{
data_test <- prepare_data(mydata)
model_object <- run_model(
 model_data = data_test,
 model_name = "constant_foi_bi",
 n_iters = 1000
)
plot_seroprev_fitted(model_object, size_text = 15)
}
}
