% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_seroprev_fitted}
\alias{plot_seroprev_fitted}
\title{Generate sero-positivity plot with fitted model}
\usage{
plot_seroprev_fitted(model_object, size_text = 6)
}
\arguments{
\item{model_object}{Object containing the results of fitting a model by means of \link{run_seroprev_model}.}

\item{size_text}{Text size of the graph returned by the function.}
}
\value{
A ggplot object containing the seropositivity-vs-age graph including the data, the fitted model and their corresponding confindence intervals.
}
\description{
Function that generates the seropositivity graph with fitted model. Age is located on the x axis and seropositivity on the y axis with its corresponding confidence interval.
}
\examples{
\dontrun{
 data_test <- prepare_seroprev_data(serodata)
 model_object <- run_seroprev_model(
 seroprev_data = data_test,
 seroprev_model_name = "constant_foi_bi",
 n_iters = 1000
)
plot_seroprev_fitted(model_object, size_text = 15)
}
}
