% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_seroprev_fitted}
\alias{plot_seroprev_fitted}
\title{Generate seropositivity plot corresponding to the specified
fitted serological model}
\usage{
plot_seroprev_fitted(
  seromodel_object,
  serodata,
  size_text = 6,
  bin_data = TRUE,
  bin_step = 5,
  alpha = 0.05
)
}
\arguments{
\item{seromodel_object}{Stanfit object containing the results of fitting a
model by means of \link{fit_seromodel}.}

\item{serodata}{A data frame containing the data from a seroprevalence
survey. This data frame must contain at least the following columns:
\describe{
\item{\code{total}}{Number of samples for each age group}
\item{\code{counts}}{Number of positive samples for each age group}
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{age_mean_f}}{Floor value of the average between age_min and age_max}
\item{\code{sample_size}}{The size of the sample}
\item{\code{birth_year}}{The year in which the individuals of each age group
were born}
}
The last six columns can be added to \code{serodata} by means of the function
\code{\link[=prepare_serodata]{prepare_serodata()}}.}

\item{size_text}{Text size of the graph returned by the function.}

\item{bin_data}{If \code{TRUE}, \code{serodata} is binned by means of
\code{prepare_bin_data}. Otherwise, age groups are kept as originally input.}

\item{bin_step}{Integer specifying the age groups bin size to be used when
\code{bin_data} is set to \code{TRUE}.}

\item{alpha}{Probability threshold for statistical significance used for both
the binomial confidence interval, and the lower and upper quantiles of the
estimated prevalence.}
}
\value{
A ggplot object containing the seropositivity-vs-age graph including
the data, the fitted model and their corresponding confidence intervals.
}
\description{
This function generates a seropositivity plot of the specified serological
model object. This includes the original data grouped by age as well as the
obtained fitting from the model implementation. Age is located on the x axis
and seropositivity on the y axis with its corresponding confidence interval.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_object <- fit_seromodel(
  serodata = serodata,
  foi_model = "constant",
  iter = 1000
)
plot_seroprev_fitted(seromodel_object,
  serodata = serodata,
  size_text = 15
)
}
