% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_seroprev_models_grid}
\alias{plot_seroprev_models_grid}
\title{Function that generates a plot showing the results for several models.}
\usage{
plot_seroprev_models_grid(..., n_row = NULL, n_col = NULL)
}
\arguments{
\item{...}{The first n arguments of the function are taken as plots ti be arranged in a single plot.
This objects can be obtained by means of \link{run_seroprev_model}.}

\item{n_row}{Number of rows of the plot array.}

\item{n_col}{Number of columns of the plot array.}
}
\value{
A ggplot object containing an array with the plots in \code{models_list}.
}
\description{
Function that generates a plot showing the results for several models.
}
\examples{
\dontrun{
 data_test <- prepare_seroprev_data(serodata)
 model_object_constant <- run_seroprev_model(seroprev_data = data_test,
                                    seroprev_model_name = "constant_foi_bi")
 model_object_normalbi <- run_seroprev_model(seroprev_data = data_test,
                                    seroprev_model_name = "continuous_foi_normal_bi")
 model_object_normallog <- run_seroprev_model(seroprev_data = data_test,
                                     seroprev_model_name = "continuous_foi_normal_log")
 plot_seroprev_models_list(model_object_constant, model_object_normalbi, model_object_normallog, n_row = 1, n_col = 3)
}
}
