% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{prepare_bin_data}
\alias{prepare_bin_data}
\title{Function that prepares a pre-processed serological survey dataset to plot the binomial confidence intervals of the seroprevalence grouped by
age group}
\usage{
prepare_bin_data(serodata)
}
\arguments{
\item{serodata}{A data frame containing the data from a seroprevalence survey. For more information see the function \link{run_seromodel}.
This data frame must contain the following columns:
\tabular{ll}{
\code{survey} \tab survey Label of the current survey \cr \tab \cr
\code{total} \tab Number of samples for each age group\cr \tab \cr
\code{counts} \tab Number of positive samples for each age group\cr \tab \cr
\code{age_min} \tab age_min \cr \tab \cr
\code{age_max} \tab age_max \cr \tab \cr
\code{tsur} \tab Year in which the survey took place \cr \tab \cr
\code{country} \tab The country where the survey took place \cr \tab \cr
\code{test} \tab The type of test taken \cr \tab \cr
\code{antibody} \tab antibody \cr \tab \cr
\code{age_mean_f} \tab Floor value of the average between age_min and age_max \cr \tab \cr
\code{sample_size} \tab The size of the sample \cr \tab \cr
\code{birth_year} \tab Years in which the subjects were borned 
according to the age group marker \code{age_mean_f}\cr \tab \cr
\code{prev_obs} \tab Observed prevalence \cr \tab \cr
\code{prev_obs_lower} \tab Lower limit of the confidence interval for the observed prevalence \cr \tab \cr
\code{prev_obs_upper} \tab Upper limit of the confidence interval for the observed prevalence \cr \tab \cr
}
The last six colums can be added to \code{serodata} by means of the function \code{\link{prepare_serodata}}.}
}
\value{
data set with the binomial confidence intervals
}
\description{
This function prepapares a given pre-processed serological dataset (see \code{\link{prepare_serodata}}) to plot the binomial confidence intervals 
of its corresponding seroprevalence grouped by age group.
}
