% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{prepare_bin_data}
\alias{prepare_bin_data}
\title{Prepare data to plot binomial confidence intervals}
\usage{
prepare_bin_data(model_data)
}
\arguments{
\item{model_data}{A data frame containing the data from a seroprevalence survey. For more information see the function \link{run_model}.}
}
\value{
data set with the binomial confidence intervals
}
\description{
Function that prepares the data for modelling
}
\examples{
\dontrun{
prepare_bin_data (model_data)
}
}
