% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{prepare_serodata}
\alias{prepare_serodata}
\title{Function that prepares the data from a serological survey for modelling}
\usage{
prepare_serodata(serodata = serodata, alpha = 0.05, add_age_mean_f = TRUE)
}
\arguments{
\item{serodata}{A data frame containing the data from a serological survey.
This data frame must contain the following columns:
\tabular{ll}{
\code{survey} \tab survey Label of the current survey \cr \tab \cr
\code{total} \tab Number of samples for each age group\cr \tab \cr
\code{counts} \tab Number of positive samples for each age group\cr \tab \cr
\code{age_min} \tab age_min \cr \tab \cr
\code{age_max} \tab age_max \cr \tab \cr
\code{tsur} \tab Year in which the survey took place \cr \tab \cr
\code{country} \tab The country where the survey took place \cr \tab \cr
\code{test} \tab The type of test taken \cr \tab \cr
\code{antibody} \tab antibody \cr \tab \cr
}}

\item{alpha}{probability of a type I error. For further details refer to \link[Hmisc]{binconf}.}

\item{add_age_mean_f}{TBD}
}
\value{
serodata with additional columns necessary for the analysis. These columns are:
\tabular{ll}{
\code{age_mean_f} \tab Floor value of the average between age_min and age_max \cr \tab \cr
\code{sample_size} \tab The size of the sample \cr \tab \cr
\code{birth_year} \tab The year in which the individuals of each age group were bornt \cr \tab \cr
\code{prev_obs} \tab Observed prevalence \cr \tab \cr
\code{prev_obs_lower} \tab Lower limit of the confidence interval for the observed prevalence \cr \tab \cr
\code{prev_obs_upper} \tab Upper limit of the confidence interval for the observed prevalence \cr \tab \cr
}
}
\description{
This function adds the necessary additional variables to the given dataset \code{serodata} corresponding to a serological survey.
}
\examples{
\dontrun{
data("serodata")
data_test <- prepare_serodata(serodata)
}
}
