% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_serosurvey.R
\name{prob_seroprev_by_age}
\alias{prob_seroprev_by_age}
\title{Generate probabilities of seropositivity by age based on model choice.}
\usage{
prob_seroprev_by_age(model, foi, seroreversion_rate = 0)
}
\arguments{
\item{model}{A string specifying the model type which can be one of
\link{'age', 'time', 'age-time'}.}

\item{foi}{A dataframe containing the force of infection (FOI) values.
For time-varying models the columns should be \link{'year', 'foi'}.
For age-varying models the columns should be \link{'age', 'foi'}.
For age-and-time-varying models the columns should be
\link{'age', 'time', 'foi'}.}

\item{seroreversion_rate}{A non-negative value determining the rate of
seroreversion (per year). Default is 0.}
}
\value{
A dataframe with columns 'age' and 'seropositivity'.
}
\description{
This function generates seropositivity probabilities based on either a
time-varying FOI model, an age-varying FOI model, or an age-and-time-varying
FOI model. In all cases, it is possible to optionally include seroreversion.
}
\examples{
prob_seroprev_by_age(
  model = "age",
  foi = data.frame(
    age = 1:80,
    foi = rep(0.01, 80)
  )
)
}
