% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_serosurvey.R
\name{probability_exact_age_varying}
\alias{probability_exact_age_varying}
\title{Computes the probability of being seropositive when FOIs vary by age}
\usage{
probability_exact_age_varying(ages, fois, seroreversion_rate = 0)
}
\arguments{
\item{ages}{Integer indicating the ages of the exposed cohorts}

\item{fois}{Numeric atomic vector corresponding to the age-varying
force-of-infection to simulate from}

\item{seroreversion_rate}{Non-negative seroreversion rate. Default is 0.}
}
\value{
vector of probabilities of being seropositive for age-varying FoI
including seroreversion (ordered from youngest to oldest individuals)
}
\description{
Computes the probability of being seropositive when FOIs vary by age
}
