% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_data.R
\name{probability_exact_age_varying}
\alias{probability_exact_age_varying}
\title{Returns the probability of being seropositive for age-varying
force-of-infection including seroreversion}
\usage{
probability_exact_age_varying(age, foi, mu = 0)
}
\arguments{
\item{age}{Integer corresponding to the age of the exposed cohort}

\item{foi}{Numeric atomic vector corresponding to the age-varying FoI to
simulate from}

\item{mu}{Seroreversion rate}
}
\value{
probability of being seropositive for age-varying FoI
including seroreversion
}
\description{
Returns the probability of being seropositive for age-varying
force-of-infection including seroreversion
}
