% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_serosurvey.R
\name{probability_exact_time_varying}
\alias{probability_exact_time_varying}
\title{Computes the probability of being seropositive when FOIs vary by time}
\usage{
probability_exact_time_varying(years, fois, seroreversion_rate = 0)
}
\arguments{
\item{years}{Integer indicating the years covering the birth ages of the
sample}

\item{fois}{Numeric atomic vector corresponding to the age-varying
force-of-infection to simulate from}

\item{seroreversion_rate}{Non-negative seroreversion rate. Default is 0.}
}
\value{
vector of probabilities of being seropositive for age-varying FoI
including seroreversion (ordered from youngest to oldest individuals)
}
\description{
Computes the probability of being seropositive when FOIs vary by time
}
