% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{run_model}
\alias{run_model}
\title{Función que corre el modelo especificado
Function that runs the specified model}
\usage{
run_model(
  model_data,
  model_name = "constant_foi_Bi",
  n_iters = 1000,
  n_thin = 2,
  delta = 0.9,
  m_treed = 10,
  decades = 0
)
}
\arguments{
\item{model_data}{model_data}

\item{model_name}{name of the model}

\item{n_iters}{number of iterations}

\item{survey}{survey}

\item{model}{model}
}
\value{
model_objects of model 0
}
\description{
Función que corre el modelo especificado
Function that runs the specified model
}
