% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{run_model}
\alias{run_model}
\title{Run model
Runs a pre-specified stan model for the force-of-infection}
\usage{
run_model(
  model_data,
  model_name = "constant_foi_bi",
  n_iters = 1000,
  n_thin = 2,
  delta = 0.9,
  m_treed = 10,
  decades = 0
)
}
\arguments{
\item{model_data}{a dataframe after using <prepare_data> function}

\item{model_name}{name of the model selected. Current version provides three options:
<constant_foi_bi>, <continuous_foi_normal_bi> and <continuous_foi_normal_log>}

\item{n_iters}{number of iterations. There is a default value by model type.}
}
\value{
model_object which is a list of various objects including stan objects and others
}
\description{
Run model
Runs a pre-specified stan model for the force-of-infection
}
