% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{run_seromodel}
\alias{run_seromodel}
\title{Function that runs the specified stan model for the Force-of-Infection and
estimates the seroprevalence based on the result of the fit}
\usage{
run_seromodel(
  serodata,
  foi_model = c("constant", "tv_normal_log", "tv_normal"),
  iter = 1000,
  thin = 2,
  adapt_delta = 0.9,
  max_treedepth = 10,
  chains = 4,
  seed = "12345",
  print_summary = TRUE,
  ...
)
}
\arguments{
\item{serodata}{A data frame containing the data from a seroprevalence
survey. This data frame must contain at least the following columns:
\describe{
\item{\code{total}}{Number of samples for each age group}
\item{\code{counts}}{Number of positive samples for each age group}
\item{\code{tsur}}{Year in which the survey took place}
\item{\code{age_mean_f}}{Floor value of the average between age_min and age_max}
\item{\code{sample_size}}{The size of the sample}
\item{\code{birth_year}}{The year in which the individuals of each age group
were born}
}
The last six columns can be added to \code{serodata} by means of the function
\code{\link[=prepare_serodata]{prepare_serodata()}}.}

\item{foi_model}{Name of the selected model. Current version provides three
options:
\describe{
\item{\code{"constant"}}{Runs a constant model}
\item{\code{"tv_normal"}}{Runs a normal model}
\item{\code{"tv_normal_log"}}{Runs a normal logarithmic model}
}}

\item{iter}{Number of interactions for each chain including the warmup.
\code{iter} in \link[rstan:stanmodel-method-sampling]{sampling}.}

\item{thin}{Positive integer specifying the period for saving samples.
\code{thin} in \link[rstan:stanmodel-method-sampling]{sampling}.}

\item{adapt_delta}{Real number between 0 and 1 that represents the target
average acceptance probability. Increasing the value of \code{adapt_delta} will
result in a smaller step size and fewer divergences. For further details
refer to the \code{control} parameter in \link[rstan:stanmodel-method-sampling]{sampling} or
\href{https://mc-stan.org/rstanarm/reference/adapt_delta.html}{here}.}

\item{max_treedepth}{Maximum tree depth for the binary tree used in the NUTS
stan sampler. For further details refer to the \code{control} parameter in
\link[rstan:stanmodel-method-sampling]{sampling}.}

\item{chains}{Number of Markov chains for sampling. For further details refer
to the \code{chains} parameter in \link[rstan:stanmodel-method-sampling]{sampling}.}

\item{seed}{For further details refer to the \code{seed} parameter in
\link[rstan:stanmodel-method-sampling]{sampling}.}

\item{print_summary}{Boolean. If \code{TRUE}, a table summarizing modelling
results is printed.}

\item{...}{Additional parameters for \link[rstan:stanmodel-method-sampling]{sampling}.}
}
\value{
\code{seromodel_object}. An object containing relevant information about
the implementation of the model. For further details refer to
\link{fit_seromodel}.
}
\description{
This function runs the specified model for the Force-of-Infection \code{foi_model}
using the data from a seroprevalence survey \code{serodata} as the input data. See
\link{fit_seromodel} for further details.
}
\examples{
data(chagas2012)
serodata <- prepare_serodata(chagas2012)
run_seromodel(
 serodata,
 foi_model = "constant"
)
}
