% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{save_or_read_model}
\alias{save_or_read_model}
\title{Save or read model
Function that saves the .RDS file of the model}
\usage{
save_or_read_model(model_name = "constant_foi_bi")
}
\arguments{
\item{model_name}{name of the model selected}
}
\description{
Save or read model
Function that saves the .RDS file of the model
}
