% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serofoi_package.R
\docType{package}
\name{serofoi-package}
\alias{serofoi}
\alias{serofoi-package}
\title{serofoi: Estimates the Force-of-Infection of a given pathogen from population based sero-prevalence studies}
\description{
R package to estimate time-varying Force-of-Infection of a given pathogen from population based sero-prevalence studies using a bayesian framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://trace-lac.github.io/serofoi/}
}

}
\author{
\strong{Maintainer}: Zulma M. Cucunubá \email{zulma.cucunuba@javeriana.edu.co}

Authors:
\itemize{
  \item Ben Lambert
  \item Pierre Nouvellet
  \item N. T. Domínguez
}

}
\keyword{internal}
