% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serofoi-package.R
\docType{package}
\name{serofoi-package}
\alias{serofoi-package}
\alias{serofoi}
\title{The 'serofoi' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R
package version 2.26.22. https://mc-stan.org

#' @keywords internal
}
\author{
\strong{Maintainer}: Zulma M. Cucunubá \email{zulma.cucunuba@javeriana.edu.co} (\href{https://orcid.org/0000-0002-8165-3198}{ORCID})

Authors:
\itemize{
  \item Nicolás T. Domínguez \email{ex-ntorres@javeriana.edu.co} (\href{https://orcid.org/0009-0002-8484-1298}{ORCID})
  \item Ben Lambert
  \item Pierre Nouvellet
}

Other contributors:
\itemize{
  \item Miguel Gámez [contributor]
  \item Geraldine Gómez (\href{https://orcid.org/0009-0007-8701-0568}{ORCID}) [contributor]
  \item Jaime A. Pavlich-Mariscal (\href{https://orcid.org/0000-0002-3892-6680}{ORCID}) [contributor]
}

}
