% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serofoi-package.R
\docType{package}
\name{serofoi-package}
\alias{serofoi-package}
\alias{serofoi}
\title{The 'serofoi' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R
package version 2.26.22. https://mc-stan.org
}
