% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_seromodel.R
\name{set_foi_init}
\alias{set_foi_init}
\title{Sets initialization function for sampling}
\usage{
set_foi_init(foi_init, is_log_foi, foi_index)
}
\arguments{
\item{foi_init}{Initialization function for sampling. If null, default is
chosen depending on the foi-scale of the model}

\item{is_log_foi}{Boolean to set logarithmic scale in the FOI}

\item{foi_index}{Integer vector specifying the age-groups for which
force-of-infection values will be estimated. It can be specified by
means of \link{get_foi_index}}
}
\description{
Sets initialization function for sampling
}
