% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stan_data.R
\name{set_stan_data_defaults}
\alias{set_stan_data_defaults}
\title{Set stan data defaults for sampling}
\usage{
set_stan_data_defaults(stan_data, is_seroreversion = FALSE)
}
\arguments{
\item{stan_data}{List to be passed to \link[rstan:stanmodel-method-sampling]{rstan}}

\item{is_seroreversion}{Boolean specifying whether to include
seroreversion rate estimation in the model}
}
\value{
List with default values of stan data for sampling
}
\description{
Set stan data defaults for sampling
}
