% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stan_data.R
\name{sf_cauchy}
\alias{sf_cauchy}
\title{Sets Cauchy distribution parameters for sampling}
\usage{
sf_cauchy(location = 0, scale = 1)
}
\arguments{
\item{location}{Location of the Cauchy distribution}

\item{scale}{Scale of the Cauchy distribution}
}
\value{
List with specified statistics and name of the distribution
}
\description{
Sets Cauchy distribution parameters for sampling
}
\examples{
my_prior <- sf_cauchy()
}
