% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata_constant.R
\docType{data}
\name{simdata_constant}
\alias{simdata_constant}
\title{Constant FoI simulated serosurvey}
\format{
An object of class \code{"cross"}; see \code{\link[qtl:read.cross]{qtl::read.cross()}}.
}
\usage{
simdata_constant
}
\description{
Simulated dataset describing a endemic situation where the Force-of-Infection
is constant over a period of 50 years (2000-2050) with a value of 0.2.
The hypothetical serosurvey is conducted in year 2050 for 250 individuals
that are up to 50 years old.
}
\examples{
simdata_constant
}
\keyword{datasets}
