% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_serosurvey.R
\name{simulate_serosurvey_general_model}
\alias{simulate_serosurvey_general_model}
\title{Simulate serosurvey data based on general serocatalytic model.}
\usage{
simulate_serosurvey_general_model(
  construct_A_function,
  calculate_seropositivity_function,
  initial_conditions,
  survey_features,
  ...
)
}
\arguments{
\item{calculate_seropositivity_function}{A function which takes the state vector
and returns the seropositive fraction.}

\item{initial_conditions}{The initial state vector proportions for each birth cohort.}

\item{survey_features}{A dataframe containing information about the binned age groups and sample
sizes for each. It should contain columns: \link{'age_min', 'age_max', 'sample_size'}.}
}
\value{
A dataframe with simulated serosurvey data, including age group information, overall
sample sizes, the number of seropositive individuals, and other survey features.
}
\description{
This simulation method assumes only that the model system can be written as a piecewise-
linear ordinary differential equation system.
}
