% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veev2012.R
\docType{data}
\name{veev2012}
\alias{veev2012}
\title{Seroprevalence data on serofoi}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
veev2012
}
\description{
Data from a serological surveys
}
\examples{
veev2012
}
\keyword{datasets}
